refactor(tui): split app.go below 1000 lines
Move Update branches, View, and key maps into sibling files
in package tui. No behavior change.

Related: #652
