+++ internal/tui/app.go
@@ -1,34 +1,18 @@
 package tui
 
 import (
-	"context"
-	"errors"
-	"fmt"
-	"image"
-	"io"
 	"os"
-	"slices"
-	"strings"
 	"time"
 
-	"charm.land/bubbles/v2/key"
 	"charm.land/bubbles/v2/spinner"
 	tea "charm.land/bubbletea/v2"
-	lipgloss "charm.land/lipgloss/v2"
-	uv "github.com/charmbracelet/ultraviolet"
-
 	"github.com/douglasdemoura/chroncal/internal/account"
 	"github.com/douglasdemoura/chroncal/internal/app"
 	"github.com/douglasdemoura/chroncal/internal/auth"
-	"github.com/douglasdemoura/chroncal/internal/caldav"
-	"github.com/douglasdemoura/chroncal/internal/calendar"
 	"github.com/douglasdemoura/chroncal/internal/config"
 	"github.com/douglasdemoura/chroncal/internal/event"
-	"github.com/douglasdemoura/chroncal/internal/icaltransfer"
 	syncpkg "github.com/douglasdemoura/chroncal/internal/sync"
-	"github.com/douglasdemoura/chroncal/internal/textsafe"
 	"github.com/douglasdemoura/chroncal/internal/trash"
-	"github.com/douglasdemoura/chroncal/internal/tui/oklch"
 )
 
 type appFocus int
@@ -46,46 +30,6 @@ const (
 	viewDay
 	viewAgenda
 )
-
-type appKeyMap struct {
-	Quit         key.Binding
-	MonthView    key.Binding
-	WeekView     key.Binding
-	DayView      key.Binding
-	AgendaView   key.Binding
-	Sidebar      key.Binding
-	WeekNumbers  key.Binding
-	WeekStart    key.Binding
-	Create       key.Binding
-	SwitchFocus  key.Binding
-	Help         key.Binding
-	Palette      key.Binding
-	CalendarList key.Binding
-	Sync         key.Binding
-	Undo         key.Binding
-	TrashView    key.Binding
-}
-
-func defaultAppKeys() appKeyMap {
-	return appKeyMap{
-		Quit:         key.NewBinding(key.WithKeys("q"), key.WithHelp("q", "quit")),
-		MonthView:    key.NewBinding(key.WithKeys("m"), key.WithHelp("m", "month")),
-		WeekView:     key.NewBinding(key.WithKeys("w"), key.WithHelp("w", "week")),
-		DayView:      key.NewBinding(key.WithKeys("d"), key.WithHelp("d", "day")),
-		AgendaView:   key.NewBinding(key.WithKeys("a"), key.WithHelp("a", "agenda")),
-		Sidebar:      key.NewBinding(key.WithKeys("\\"), key.WithHelp("\\", "sidebar")),
-		WeekNumbers:  key.NewBinding(key.WithKeys("#"), key.WithHelp("#", "week numbers")),
-		WeekStart:    key.NewBinding(key.WithKeys("W", "shift+w"), key.WithHelp("W", "week start")),
-		Create:       key.NewBinding(key.WithKeys("c"), key.WithHelp("c", "new")),
-		SwitchFocus:  key.NewBinding(key.WithKeys("tab", "shift+tab"), key.WithHelp("tab", "switch focus")),
-		Help:         key.NewBinding(key.WithKeys("?"), key.WithHelp("?", "help")),
-		Palette:      key.NewBinding(key.WithKeys("/", "ctrl+k"), key.WithHelp("/", "commands")),
-		CalendarList: key.NewBinding(key.WithKeys("C", "shift+c"), key.WithHelp("C", "calendars")),
-		Sync:         key.NewBinding(key.WithKeys("s"), key.WithHelp("s", "sync")),
-		Undo:         key.NewBinding(key.WithKeys("u"), key.WithHelp("u", "undo")),
-		TrashView:    key.NewBinding(key.WithKeys("D", "shift+d"), key.WithHelp("D", "recently deleted")),
-	}
-}
 
 type eventsLoadedMsg struct {
 	// from and to identify the query range. The handler can then drop stale
@@ -627,174 +571,6 @@ func newAgendaForStartup(now time.Time, vm viewMode, showEmptyDays bool) AgendaM
 	return a
 }
 
-// expectedEventRange returns the [from, to) UTC range the active view
-// currently expects from loadEvents. It seeds each query. It also
-// validates eventsLoadedMsg that arrive against stale async responses.
-func (m Model) expectedEventRange() (time.Time, time.Time) {
-	switch m.viewMode {
-	case viewDay:
-		d := m.day.Cursor()
-		return localSpanQueryRange(d, d.AddDate(0, 0, 1))
-	case viewWeek:
-		start := m.week.WeekStartDate()
-		return localSpanQueryRange(start, start.AddDate(0, 0, 7))
-	case viewAgenda:
-		return localSpanQueryRange(m.agenda.WindowStart(), m.agenda.WindowEnd())
-	default:
-		anchor := calendarGridAnchor(m.calendar.Month(), m.calendar.WeekStart())
-		return localSpanQueryRange(anchor, anchor.AddDate(0, 0, 42))
-	}
-}
-
-// localSpanQueryRange returns the [from, to) UTC query range that covers
-// the local calendar days [fromDay, toDay). The range is the union of two
-// spans. The UTC-midnight span covers the all-day events. The database
-// stores those as 00:00 UTC datestamps. The local-midnight span, converted
-// to UTC, covers the timed events. The database stores those as UTC
-// instants, so an evening event in a UTC-negative zone lands on the next
-// UTC date. The UTC-midnight span alone misses that event (the day view
-// then hides it while the wider views show it).
-func localSpanQueryRange(fromDay, toDay time.Time) (time.Time, time.Time) {
-	from := time.Date(fromDay.Year(), fromDay.Month(), fromDay.Day(), 0, 0, 0, 0, time.UTC)
-	to := time.Date(toDay.Year(), toDay.Month(), toDay.Day(), 0, 0, 0, 0, time.UTC)
-	localFrom := time.Date(fromDay.Year(), fromDay.Month(), fromDay.Day(), 0, 0, 0, 0, time.Local).UTC()
-	localTo := time.Date(toDay.Year(), toDay.Month(), toDay.Day(), 0, 0, 0, 0, time.Local).UTC()
-	if localFrom.Before(from) {
-		from = localFrom
-	}
-	if localTo.After(to) {
-		to = localTo
-	}
-	return from, to
-}
-
-// dispatchEditScope routes a deferred EventFormSaveMsg to the right service
-// method based on the user's choice in the scope dialog. choice indices match
-// the dialog order: 0=this event only, 1=this and following, 2=all events.
-// Returns the message the agenda loop should consume next.
-func (m Model) dispatchEditScope(editID int64, choice int, save EventFormSaveMsg) tea.Msg {
-	ctx := context.Background()
-	master, err := m.app.Events.Get(ctx, editID)
-	if err != nil {
-		return eventUpdateAfterScopeMsg{calendarID: save.CalendarID, err: err}
-	}
-
-	params := event.UpdateParams{
-		CalendarID:     save.CalendarID,
-		Title:          save.Title,
-		Description:    save.Description,
-		Location:       save.Location,
-		ConferenceURI:  save.ConferenceURI,
-		StartTime:      save.StartTime,
-		EndTime:        save.EndTime,
-		AllDay:         save.AllDay,
-		RecurrenceRule: save.RecurrenceRule,
-		Timezone:       save.Timezone,
-		Transp:         save.Transp,
-		Class:          save.Class,
-		Categories:     save.Categories,
-	}
-
-	// Each branch writes the event row together with its attendees and alarms
-	// in a single transaction. A failed child write then rolls the whole edit
-	// back. It does not leave a half-updated row (issue #87).
-	switch choice {
-	case 0: // This event only
-		_, err = m.app.Events.UpdateInstanceWithRelations(ctx, master.UID, save.InstanceTime, params, save.Attendees, save.Alarms)
-	case 1: // This and following
-		_, err = m.app.Events.UpdateFromInstanceWithRelations(ctx, master.UID, save.InstanceTime, params, save.Attendees, save.Alarms)
-	default: // All events
-		// The form opened showing the clicked instance's time, so save.StartTime
-		// is relative to that occurrence — not the master's DTSTART. Apply the
-		// delta to the master to shift the whole series, and preserve whatever
-		// duration the user left on the form.
-		delta := save.StartTime.Sub(save.InstanceTime)
-		newDuration := save.EndTime.Sub(save.StartTime)
-		params.StartTime = master.StartTime.Add(delta)
-		params.EndTime = params.StartTime.Add(newDuration)
-		_, err = m.app.Events.UpdateWithRelations(ctx, master.ID, params, save.Attendees, save.Alarms)
-	}
-	if err != nil {
-		return eventUpdateAfterScopeMsg{calendarID: save.CalendarID, err: err}
-	}
-	return eventUpdateAfterScopeMsg{calendarID: save.CalendarID}
-}
-
-func (m Model) loadEvents() tea.Cmd {
-	from, to := m.expectedEventRange()
-	return m.queryEventsRange(from, to, false)
-}
-
-// loadEventsIncremental queries only the newly-added slice of an agenda
-// expansion when the loaded range shares an edge with the new expected
-// range. Infinite-scroll then stays O(1 step) in query cost even after the
-// user has scrolled years back. It falls back to a full refresh when the
-// ranges do not share an edge (for example after a cursor jump).
-func (m Model) loadEventsIncremental() tea.Cmd {
-	wantFrom, wantTo := m.expectedEventRange()
-	if m.loadedFrom.IsZero() || m.loadedTo.IsZero() {
-		return m.queryEventsRange(wantFrom, wantTo, false)
-	}
-	// Forward extension: near edge unchanged, far edge pushed later.
-	if m.loadedFrom.Equal(wantFrom) && m.loadedTo.Before(wantTo) {
-		return m.queryEventsRange(m.loadedTo, wantTo, true)
-	}
-	// Backward extension: far edge unchanged, near edge pushed earlier.
-	if m.loadedTo.Equal(wantTo) && wantFrom.Before(m.loadedFrom) {
-		return m.queryEventsRange(wantFrom, m.loadedFrom, true)
-	}
-	// No shared edge — full refresh.
-	return m.queryEventsRange(wantFrom, wantTo, false)
-}
-
-// queryEventsRange runs the recurrence-expanded query for [from, to).
-// It returns an eventsLoadedMsg tagged with the queried range. The tag
-// also says whether the result is a merge (incremental) or a replacement
-// (full refresh).
-func (m Model) queryEventsRange(from, to time.Time, merge bool) tea.Cmd {
-	return func() tea.Msg {
-		expanded, err := m.app.Recurrences.ListExpandedEvents(context.Background(), from, to)
-		events := make([]event.Event, len(expanded))
-		for i, e := range expanded {
-			evt := e.Event
-			if !evt.EndTime.IsZero() {
-				evt.EndTime = e.InstanceTime.Add(evt.EndTime.Sub(evt.StartTime))
-			}
-			evt.StartTime = e.InstanceTime
-			events[i] = evt
-		}
-		return eventsLoadedMsg{from: from, to: to, merge: merge, events: events, err: err}
-	}
-}
-
-// mergeEvents dedup-appends new events into the stored list. The dedup key is
-// (ID, StartTime.UTC()). That key is unique for both non-recurring events and
-// recurrence instances. Needed when a multi-day event straddles the
-// incremental slice boundary and is returned by both queries.
-func mergeEvents(existing, incoming []event.Event) []event.Event {
-	seen := make(map[string]struct{}, len(existing)+len(incoming))
-	out := make([]event.Event, 0, len(existing)+len(incoming))
-	add := func(e event.Event) {
-		key := eventDedupKey(e)
-		if _, ok := seen[key]; ok {
-			return
-		}
-		seen[key] = struct{}{}
-		out = append(out, e)
-	}
-	for _, e := range existing {
-		add(e)
-	}
-	for _, e := range incoming {
-		add(e)
-	}
-	return out
-}
-
-func eventDedupKey(e event.Event) string {
-	return e.StartTime.UTC().Format(time.RFC3339) + "|" + fmt.Sprint(e.ID)
-}
-
 // trashLoadedMsg carries trash entries across events, todos, and journals
 // for the visible calendar(s), plus an error if any domain query failed.
 type trashLoadedMsg struct {
@@ -810,4578 +586,12 @@ type trashActionDoneMsg struct {
 	err    error
 }
 
-// loadTrash queries the trash aggregator across all visible calendars
-// and hands the result to the trash model via trashLoadedMsg.
-func (m Model) loadTrash() tea.Cmd {
-	return func() tea.Msg {
-		ctx := context.Background()
-		var out []trash.Entry
-		for id := range m.calendars {
-			if m.hiddenCalendars[id] {
-				continue
-			}
-			entries, err := m.app.Trash.List(ctx, id)
-			if err != nil {
-				return trashLoadedMsg{err: err}
-			}
-			out = append(out, entries...)
-		}
-		return trashLoadedMsg{entries: out}
-	}
-}
-
-func eventsOn(events []event.Event, day time.Time) []event.Event {
-	dayKey := day.Local().Format("2006-01-02")
-	var out []event.Event
-	for _, e := range events {
-		// All-day events are stored as midnight UTC; compare in UTC
-		// so negative-offset timezones don't shift the date.
-		eKey := e.StartTime.Local().Format("2006-01-02")
-		if e.AllDay {
-			eKey = e.StartTime.UTC().Format("2006-01-02")
-		}
-		if eKey == dayKey {
-			out = append(out, e)
-		}
-	}
-	return out
-}
-
-func eventsToCalendar(events []event.Event, calendars map[int64]CalendarInfo, hidden map[int64]bool) []CalendarEvent {
-	out := make([]CalendarEvent, 0, len(events))
-	for _, e := range events {
-		if hidden[e.CalendarID] {
-			continue
-		}
-		color := calendars[e.CalendarID].Color
-		for _, day := range eventCalendarDays(e) {
-			start, end := clipEventToDay(e, day)
-			out = append(out, CalendarEvent{
-				ID:        e.ID,
-				Title:     e.Title,
-				AllDay:    e.AllDay,
-				Day:       day,
-				Color:     color,
-				StartTime: start,
-				EndTime:   end,
-			})
-		}
-	}
-	return out
-}
-
-// eventCalendarDays returns one entry for each local calendar day an event
-// touches. All-day events use UTC (their StartTime is a datestamp at 00:00
-// UTC, not a point in time). Timed events use local time.
-func eventCalendarDays(e event.Event) []time.Time {
-	if e.AllDay {
-		s := e.StartTime.UTC()
-		startDay := time.Date(s.Year(), s.Month(), s.Day(), 0, 0, 0, 0, time.UTC)
-		end := e.EndTime.UTC()
-		var days []time.Time
-		for d := startDay; d.Before(end); d = d.AddDate(0, 0, 1) {
-			days = append(days, d)
-		}
-		if len(days) == 0 {
-			days = []time.Time{startDay}
-		}
-		return days
-	}
-	s := e.StartTime.Local()
-	end := e.EndTime.Local()
-	startDay := time.Date(s.Year(), s.Month(), s.Day(), 0, 0, 0, 0, s.Location())
-	if !end.After(s) {
-		return []time.Time{startDay}
-	}
-	var days []time.Time
-	for d := startDay; d.Before(end); d = d.AddDate(0, 0, 1) {
-		days = append(days, d)
-	}
-	if len(days) == 0 {
-		days = []time.Time{startDay}
-	}
-	return days
-}
-
-// clipEventToDay returns the event's start and end times clipped to the
-// given calendar day. For all-day events the times are the event's original
-// values (views ignore them). For timed events that span midnight, the end
-// of day 1 is pushed one second before midnight. The time-grid renderer
-// then sees an in-day hour/minute (placeEvents reads only hour/minute).
-func clipEventToDay(e event.Event, day time.Time) (time.Time, time.Time) {
-	if e.AllDay {
-		return e.StartTime, e.EndTime
-	}
-	dayStart := time.Date(day.Year(), day.Month(), day.Day(), 0, 0, 0, 0, day.Location())
-	dayEnd := dayStart.AddDate(0, 0, 1)
-	start := e.StartTime.Local()
-	if start.Before(dayStart) {
-		start = dayStart
-	}
-	end := e.EndTime.Local()
-	if !end.Before(dayEnd) {
-		end = dayEnd.Add(-time.Second)
-	}
-	return start, end
-}
-
-func (m Model) loadCalendars() tea.Cmd {
-	return func() tea.Msg {
-		ctx := context.Background()
-		cals, err := m.app.Calendars.List(ctx)
-		if err != nil {
-			return calendarsLoadedMsg{err: err}
-		}
-		// Pre-fetch accounts once so linked rows carry their user-facing group
-		// name, server URL, normalized auth type, and order without an account
-		// query per calendar.
-		accounts, _ := m.app.Accounts.List(ctx)
-		info := buildCalendarInfoMap(cals, accounts, func(calendarID int64) (int64, error) {
-			return m.app.Events.CountByCalendar(ctx, calendarID)
-		})
-		accountByID := make(map[int64]account.Account, len(accounts))
-		for _, configured := range accounts {
-			accountByID[configured.ID] = configured
-		}
-		return calendarsLoadedMsg{calendars: info, accounts: accountByID}
-	}
-}
-
-// buildCalendarInfoMap assembles the sidebar's CalendarInfo cache from the
-// persisted calendars and accounts. Accounts are read once so every linked
-// row carries its account metadata: display name, server URL, normalized
-// auth type, and display order. There is no per-calendar query. Local
-// calendars (AccountID == 0) leave the account-linked fields empty,
-// AccountAuthType included. A nil or empty accounts slice yields local-only
-// metadata. That is the same effect as an account-list failure.
-func buildCalendarInfoMap(
-	cals []calendar.Calendar,
-	accounts []account.Account,
-	eventCount func(calendarID int64) (int64, error),
-) map[int64]CalendarInfo {
-	accountServerURLs := map[int64]string{}
-	accountNames := map[int64]string{}
-	accountOrders := map[int64]int64{}
-	accountAuthTypes := map[int64]string{}
-	for _, remoteAccount := range accounts {
-		accountServerURLs[remoteAccount.ID] = remoteAccount.ServerURL
-		accountNames[remoteAccount.ID] = remoteAccount.DisplayName
-		accountOrders[remoteAccount.ID] = remoteAccount.DisplayOrder
-		accountAuthTypes[remoteAccount.ID] = calendar.NormalizeAuthType(remoteAccount.AuthType)
-	}
-	info := make(map[int64]CalendarInfo, len(cals))
-	for _, c := range cals {
-		count, _ := eventCount(c.ID)
-		info[c.ID] = CalendarInfo{
-			Name:                c.Name,
-			Color:               c.Color,
-			OwnerEmail:          c.OwnerEmail,
-			Description:         c.Description,
-			EventCount:          count,
-			DisplayOrder:        c.DisplayOrder,
-			Synced:              c.AccountID != 0,
-			AccountServerURL:    accountServerURLs[c.AccountID],
-			AccountID:           c.AccountID,
-			AccountName:         accountNames[c.AccountID],
-			AccountOrder:        accountOrders[c.AccountID],
-			AccountAuthType:     accountAuthTypes[c.AccountID],
-			RemoteAccess:        c.RemoteAccess,
-			RemoteComponents:    c.RemoteComponents,
-			RemoteMissing:       c.RemoteMissing,
-			LastSyncAt:          c.LastSyncAt,
-			LastSyncAttemptedAt: c.LastSyncAttemptedAt,
-			LastSyncError:       c.LastSyncError,
-			CreatedAt:           c.CreatedAt,
-			UpdatedAt:           c.UpdatedAt,
-			IsDefault:           c.IsDefault,
-		}
-	}
-	return info
-}
-
-func (m Model) accountSettingsParams(accountID int64) (AccountSettingsParams, bool) {
-	configured, ok := m.accounts[accountID]
-	if !ok || accountID == 0 {
-		return AccountSettingsParams{}, false
-	}
-	params := AccountSettingsParams{
-		AccountID:   accountID,
-		DisplayName: configured.DisplayName,
-		Provider:    "CalDAV Account",
-		ServerURL:   configured.ServerURL,
-		Username:    configured.Username,
-		AuthType:    calendar.NormalizeAuthType(configured.AuthType),
-	}
-	if caldav.IsGoogleCalendarEndpoint(configured.ServerURL) {
-		params.Provider = "Google Account"
-	}
-	for _, info := range m.calendars {
-		if info.AccountID != accountID {
-			continue
-		}
-		params.CalendarCount++
-		if info.LastSyncError != "" {
-			params.AttentionCount++
-		}
-	}
-	return params, true
-}
-
-func (m Model) reopenAccountSettings(accountID int64) Model {
-	if params, ok := m.accountSettingsParams(accountID); ok {
-		m.calendarManager = m.calendarManager.OpenAccount(params).SetSize(m.width, m.height)
-		m.calendarManagerOpen = true
-	}
-	return m
-}
-
-func (m Model) finishAccountRename(msg accountRenameFinishedMsg) (Model, tea.Cmd) {
-	m.syncing = false
-	m.statusToken++
-	if msg.err != nil {
-		if m.accountRenameFromSettings {
-			m.accountRename.form.SetError(0, msg.err.Error())
-			m.accountRenameOpen = true
-		}
-		m.syncStatus = "Account rename failed: " + msg.err.Error()
-		return m, tea.Batch(
-			m.toast.Failed(msg.err.Error()),
-			m.expireStatusAfter(8*time.Second, m.statusToken),
-		)
-	}
-	m.accountRenameFromSettings = false
-	if m.accounts == nil {
-		m.accounts = make(map[int64]account.Account)
-	}
-	m.accounts[msg.account.ID] = msg.account
-	if m.calendarManagerOpen && m.calendarManager.ActiveAccountID() == msg.account.ID {
-		if params, ok := m.accountSettingsParams(msg.account.ID); ok {
-			m.calendarManager = m.calendarManager.OpenAccount(params).SetSize(m.width, m.height)
-		}
-	}
-	if draft := m.calendarManager.LocalDraft(); draft != nil &&
-		draft.AccountID == msg.account.ID {
-		m.calendarManager = m.calendarManager.SetAccountName(msg.account.DisplayName)
-	}
-	m.syncStatus = fmt.Sprintf("Renamed account to %s", msg.account.DisplayName)
-	return m, tea.Batch(
-		m.loadCalendars(),
-		m.expireStatusAfter(6*time.Second, m.statusToken),
-	)
-}
-
-func (m Model) finishAccountRemoval(msg accountRemovalFinishedMsg) (Model, tea.Cmd) {
-	m.syncing = false
-	m.statusToken++
-	if msg.err != nil {
-		if !m.calendarManagerOpen {
-			m = m.reopenAccountSettings(msg.accountID)
-		}
-		m.syncStatus = "Account removal failed: " + msg.err.Error()
-		return m, tea.Batch(
-			m.toast.Failed(msg.err.Error()),
-			m.expireStatusAfter(10*time.Second, m.statusToken),
-		)
-	}
-	if m.calendarManagerOpen {
-		if settings, ok := m.calendarManager.AccountSettings(); ok && settings.params.AccountID == msg.accountID {
-			m.calendarManagerOpen = false
-		}
-	}
-	if draft := m.calendarManager.LocalDraft(); draft != nil && draft.AccountID == msg.accountID {
-		m.calendarManagerOpen = false
-	}
-	m.syncStatus = fmt.Sprintf(
-		"Removed account %s; downloaded calendars are now local",
-		textsafe.Display(msg.name),
-	)
-	return m, tea.Batch(
-		m.loadCalendars(),
-		m.expireStatusAfter(10*time.Second, m.statusToken),
-	)
-}
-
-func (m Model) finishAccountManagementDiscovery(
-	msg accountManagementDiscoveryReadyMsg,
-) (Model, tea.Cmd) {
-	stale := msg.generation != m.accountManagementGeneration ||
-		msg.accountID != m.pendingAccountManagementID
-	if configured, ok := m.accounts[msg.accountID]; !ok || configured.ID == 0 {
-		stale = true
-	}
-	if msg.err == nil && msg.discovery.Account.ID != msg.accountID {
-		stale = true
-	}
-	m.syncing = false
-	if stale {
-		m.pendingAccountManagementID = 0
-		m.syncStatus = ""
-		return m, nil
-	}
-	if msg.err != nil {
-		m.pendingAccountManagementID = 0
-		if params, ok := m.accountSettingsParams(msg.accountID); ok {
-			m.calendarManager = m.calendarManager.OpenAccount(params).SetSize(m.width, m.height)
-			m.calendarManagerOpen = true
-		}
-		m.statusToken++
-		m.syncStatus = "Calendar discovery failed: " + msg.err.Error()
-		return m, tea.Batch(
-			m.toast.Failed(msg.err.Error()),
-			m.expireStatusAfter(10*time.Second, m.statusToken),
-		)
-	}
-	m.pendingAccountManagementID = 0
-	m.calendarManager = m.calendarManager.OpenAccountCalendars(msg.discovery).SetSize(m.width, m.height)
-	m.calendarManagerOpen = true
-	m.syncStatus = ""
-	return m, m.loadCalendars()
-}
-
-// sortedCalendarListItems builds the sidebar's calendar rows from the calendar
-// map and sorts them by the user's persisted display order (name as a tiebreak,
-// e.g. rows that share a backfilled default). Shared by the reload handler and
-// the reorder handler so both produce identical row order.
-func sortedCalendarListItems(calendars map[int64]CalendarInfo) []CalendarListItem {
-	items := make([]CalendarListItem, 0, len(calendars))
-	for id, calendarInfo := range calendars {
-		accountName := calendarInfo.AccountName
-		if calendarInfo.AccountID == 0 {
-			accountName = "Local"
-		}
-		items = append(items, CalendarListItem{
-			ID:           id,
-			Name:         calendarInfo.Name,
-			Color:        calendarInfo.Color,
-			Health:       syncHealthFor(calendarInfo),
-			Order:        calendarInfo.DisplayOrder,
-			AccountID:    calendarInfo.AccountID,
-			AccountName:  accountName,
-			AccountOrder: calendarInfo.AccountOrder,
-			Access:       calendarInfo.RemoteAccess,
-			Missing:      calendarInfo.RemoteMissing,
-		})
-	}
-	slices.SortFunc(items, func(a, b CalendarListItem) int {
-		if a.AccountID == 0 && b.AccountID != 0 {
-			return -1
-		}
-		if a.AccountID != 0 && b.AccountID == 0 {
-			return 1
-		}
-		if a.AccountID != b.AccountID {
-			if a.AccountOrder < b.AccountOrder {
-				return -1
-			}
-			if a.AccountOrder > b.AccountOrder {
-				return 1
-			}
-			if a.AccountID < b.AccountID {
-				return -1
-			}
-			return 1
-		}
-		return compareCalendarOrder(a.Order, a.Name, b.Order, b.Name)
-	})
-	return items
-}
-
-func (m *Model) beginAccountOrderSave(ids []int64) tea.Cmd {
-	saved := slices.Clone(ids)
-	m.accountOrderSaveInFlight = true
-	return func() tea.Msg {
-		return accountOrderSavedMsg{
-			ids: saved,
-			err: m.app.Accounts.SetOrder(context.Background(), saved),
-		}
-	}
-}
-
-// blockReadOnlyCalendarMutation rejects event mutations on calendars the remote
-// server has declared off-limits for VEVENT writes. Those are read-only
-// collections, and collections whose non-empty supported-component set omits
-// VEVENT. A calendar with no reported components is treated as unconstrained
-// (backward compatible). Service-layer guards are enforced separately.
-func (m *Model) blockReadOnlyCalendarMutation(calendarID int64) (tea.Cmd, bool) {
-	info, ok := m.calendars[calendarID]
-	if !ok {
-		return nil, false
-	}
-	name := strings.TrimSpace(info.Name)
-	if name == "" {
-		name = "This calendar"
-	}
-	if strings.EqualFold(strings.TrimSpace(info.RemoteAccess), "read") {
-		return m.toast.Failed(fmt.Sprintf("%s is read-only", name)), true
-	}
-	if info.RemoteComponents != "" && !supportsVEVENT(info.RemoteComponents) {
-		return m.toast.Failed(fmt.Sprintf("%s does not support events", name)), true
-	}
-	return nil, false
-}
-
-// supportsVEVENT reports whether the comma-separated component set contains
-// VEVENT. An empty string means "no advertised components" and is handled by
-// the caller, not here.
-func supportsVEVENT(remoteComponents string) bool {
-	for _, part := range strings.Split(remoteComponents, ",") {
-		if strings.EqualFold(strings.TrimSpace(part), "VEVENT") {
-			return true
-		}
-	}
-	return false
-}
-
-// eventFormCalendars omits collections that cannot accept VEVENT writes so
-// create and move pickers never offer a destination that the service must
-// reject on submit. Empty component metadata remains backward compatible.
-func eventFormCalendars(calendars map[int64]CalendarInfo) map[int64]CalendarInfo {
-	filtered := make(map[int64]CalendarInfo, len(calendars))
-	for id, info := range calendars {
-		if strings.EqualFold(strings.TrimSpace(info.RemoteAccess), "read") {
-			continue
-		}
-		if info.RemoteComponents != "" && !supportsVEVENT(info.RemoteComponents) {
-			continue
-		}
-		filtered[id] = info
-	}
-	return filtered
-}
-
-// syncHealthFor derives the sidebar health marker state from a calendar's
-// persisted sync fields. Local-only calendars (not Synced) get no marker.
-// A recorded last_sync_error is the only loud (SyncHealthError) state.
-// Every sync attempt writes these fields: manual, and the background
-// `chroncal tick` cron. The marker then surfaces failures the user never
-// triggered.
-func syncHealthFor(info CalendarInfo) SyncHealth {
-	switch {
-	case !info.Synced:
-		return SyncHealthNone
-	case info.LastSyncError != "":
-		return SyncHealthError
-	case info.LastSyncAt != "":
-		return SyncHealthOK
-	default:
-		return SyncHealthPending
-	}
-}
-
-// openCredentialStore opens the credential store with the namespace settings and
-// the migration settings of the app. It discards the warnings, so keyring output
-// never overwrites the rendered TUI. Every TUI flow that touches a credential
-// goes through this one constructor. A change to how the store opens thus
-// applies to all the flows at the same time.
-func (m Model) openCredentialStore() (auth.CredentialStore, error) {
-	return auth.NewCredentialStoreWithWarnings(
-		m.app.CredentialNamespace,
-		m.app.PreviousCredentialNamespaces,
-		m.app.MigrateLegacyCredentials,
-		m.app.AllowPlaintext,
-		io.Discard,
-	)
-}
-
-// newSyncService builds a sync.Service using the app's shared SQLite handle.
-// Credential-store warnings are discarded so sync work does not clobber the
-// rendered TUI. Token-refresh persist mid-run is included. The engine logs
-// to the state-dir log file (never stderr — Bubble Tea owns the terminal).
-// Sync detail like import warnings then stays inspectable. Users run
-// `chroncal sync run` from a shell if they need verbose output live.
-func (m Model) newSyncService() (*syncpkg.Service, error) {
-	credStore, err := m.openCredentialStore()
-	if err != nil {
-		return nil, fmt.Errorf("credential store: %w", err)
-	}
-	return syncpkg.NewService(m.app.DB, m.app.Queries, credStore, m.app.Calendars, m.app.Events, m.app.Todos, m.app.Journals, config.SharedStateDirLogger()), nil
-}
-
-// runSyncAllPlan lists the connected calendars and emits a syncAllPlannedMsg
-// so the Update loop can step through them one at a time. The actual sync work
-// happens in runSyncOne so the footer can refresh between calendars.
-func (m Model) runSyncAllPlan() tea.Cmd {
-	return func() tea.Msg {
-		ctx := context.Background()
-		cals, err := m.app.Queries.ListCalendars(ctx)
-		if err != nil {
-			return syncFinishedMsg{err: err}
-		}
-		var targets []syncTarget
-		for _, c := range cals {
-			if c.AccountID == nil || *c.AccountID == 0 {
-				continue
-			}
-			targets = append(targets, syncTarget{ID: c.ID, Name: c.Name})
-		}
-		return syncAllPlannedMsg{targets: targets}
-	}
-}
-
-// runSyncOne syncs a single calendar inside a SyncAll run. It emits
-// syncCalendarFinishedMsg so the Update loop can advance to the next target
-// (or finalize) and refresh the footer.
-func (m Model) runSyncOne(target syncTarget, index, total int) tea.Cmd {
-	return func() tea.Msg {
-		svc, err := m.newSyncService()
-		if err != nil {
-			return syncCalendarFinishedMsg{index: index, total: total, name: target.Name, err: err}
-		}
-		ctx, cancel := context.WithTimeout(context.Background(), 5*time.Minute)
-		defer cancel()
-		result, err := svc.SyncCalendar(ctx, target.ID, m.fullSyncStrategy)
-		return syncCalendarFinishedMsg{index: index, total: total, name: target.Name, result: result, err: err}
-	}
-}
-
-func (m Model) runSyncCalendar(id int64, name string) tea.Cmd {
-	return func() tea.Msg {
-		svc, err := m.newSyncService()
-		if err != nil {
-			return syncFinishedMsg{err: err}
-		}
-		ctx, cancel := context.WithTimeout(context.Background(), 5*time.Minute)
-		defer cancel()
-		result, err := svc.SyncCalendar(ctx, id, m.fullSyncStrategy)
-		if err != nil {
-			return syncFinishedMsg{err: err}
-		}
-		label := name
-		if label == "" {
-			label = "calendar"
-		}
-		summary := syncSummary(label, syncTotals{
-			pushed:           result.Pushed,
-			pulled:           result.Pulled,
-			deleted:          result.Deleted,
-			conflicts:        result.Conflicts,
-			autoResolved:     result.AutoResolved,
-			skippedConflicts: result.SkippedConflicts,
-			warnings:         len(result.Warnings),
-		})
-		var firstErr error
-		if len(result.Errors) > 0 {
-			firstErr = result.Errors[0]
-		}
-		return syncFinishedMsg{summary: summary, err: firstErr, reload: true}
-	}
-}
-
-func (m Model) runSyncAccount(accountID int64, name string) tea.Cmd {
-	return func() tea.Msg {
-		svc, err := m.newSyncService()
-		if err != nil {
-			return syncFinishedMsg{err: err}
-		}
-		results, err := svc.SyncAccount(
-			context.Background(), accountID, m.fullSyncStrategy,
-		)
-		if err != nil {
-			return syncFinishedMsg{err: err}
-		}
-		var totals syncTotals
-		for _, result := range results {
-			if result == nil {
-				continue
-			}
-			totals.pushed += result.Pushed
-			totals.pulled += result.Pulled
-			totals.deleted += result.Deleted
-			totals.conflicts += result.Conflicts
-			totals.autoResolved += result.AutoResolved
-			totals.skippedConflicts += result.SkippedConflicts
-			totals.warnings += len(result.Warnings)
-			if totals.firstErr == nil && len(result.Errors) > 0 {
-				totals.firstErr = result.Errors[0]
-			}
-		}
-		label := name
-		if label == "" {
-			label = "account"
-		}
-		return syncFinishedMsg{
-			summary: syncSummary(label, totals),
-			err:     totals.firstErr,
-			reload:  true,
-		}
-	}
-}
-
-func (m Model) finishSync(msg syncFinishedMsg) (Model, tea.Cmd) {
-	m.syncing = false
-	m.statusToken++
-	if msg.err != nil {
-		if msg.summary != "" {
-			m.syncStatus = fmt.Sprintf("%s — %s", msg.summary, msg.err.Error())
-		} else {
-			m.syncStatus = "Sync failed: " + msg.err.Error()
-		}
-	} else {
-		m.syncStatus = msg.summary
-	}
-	// Always reload calendars so the sidebar health marker reflects the
-	// just-attempted sync — including a failed run. Events only change on a
-	// successful pull, so reload those only then.
-	cmds := []tea.Cmd{m.expireStatusAfter(6*time.Second, m.statusToken), m.loadCalendars()}
-	if msg.reload {
-		cmds = append(cmds, m.loadEvents())
-	}
-	if m.pendingSyncCalendar.ID != 0 {
-		next := m.pendingSyncCalendar
-		m.pendingSyncCalendar = syncTarget{}
-		cmds = append(cmds, func() tea.Msg {
-			return SyncCalendarRequestedMsg(next)
-		})
-	}
-	return m, tea.Batch(cmds...)
-}
-
-// startOAuthFlow opens the wait modal and launches browser authorization
-// with the given client config. Add Account consumes the sign-in dialog.
-// Account re-authentication keeps any edit draft under the flow.
-// The caller has already recorded m.oauthPurpose.
-func (m Model) startOAuthFlow(clientID, clientSecret string) (Model, tea.Cmd) {
-	m.oauthPending = false // the flow is opening now; release the request guard
-	if m.oauthPurpose.calendarDiscovery {
-		m.calendarManagerOpen = false
-	}
-	m.calendarManagerOpen = false
-	m.accountOAuthConfigOpen = false
-	m.oauthFlowOpen = true
-	m.oauthFlow = m.oauthFlow.SetSize(m.width, m.height)
-	var cmd tea.Cmd
-	m.oauthFlow, cmd = m.oauthFlow.Start(clientID, clientSecret)
-	return m, cmd
-}
-
-func (m Model) prepareAccountReauth(
-	configured account.Account,
-	clientID, clientSecret string,
-) tea.Cmd {
-	return func() tea.Msg {
-		ctx := context.Background()
-		credStore, err := m.openCredentialStore()
-		if err != nil {
-			return accountReauthReadyMsg{accountID: configured.ID, name: configured.DisplayName, err: err}
-		}
-		cred, err := m.app.Accounts.LoadCredential(ctx, configured.ID, credStore)
-		if err != nil {
-			return accountReauthReadyMsg{
-				accountID: configured.ID,
-				name:      configured.DisplayName,
-				err:       fmt.Errorf("load credential: %w", err),
-			}
-		}
-		if clientID != "" {
-			cred.OAuthClientID = clientID
-		}
-		if clientSecret != "" {
-			cred.OAuthClientSecret = clientSecret
-		}
-		return accountReauthReadyMsg{
-			accountID: configured.ID,
-			name:      configured.DisplayName,
-			cred:      cred,
-		}
-	}
-}
-
-// finishOAuthReauth persists the fresh tokens for a re-authenticated
-// account. A full re-consent always returns a new refresh token. That
-// differs from RefreshGoogleToken, which only returns a new one if the
-// server rotates it. The stored credential's username and client config
-// are kept. The token triple is replaced.
-func (m Model) finishOAuthReauth(result *auth.GoogleOAuthResult) tea.Cmd {
-	p := m.oauthPurpose
-	storedMsg := func(err error) oauthCredentialStoredMsg {
-		return oauthCredentialStoredMsg{
-			accountID: p.accountID,
-			name:      p.accountName,
-			err:       err,
-		}
-	}
-	return func() tea.Msg {
-		credStore, err := m.openCredentialStore()
-		if err != nil {
-			return storedMsg(err)
-		}
-		cred := p.cred
-		cred.AccountID = p.accountID
-		cred.AccessToken = result.AccessToken
-		// Keep the current refresh token if Google did not return a new
-		// one. access_type=offline&prompt=consent normally forces a fresh
-		// refresh token. If it comes back empty, an overwrite would
-		// strip the account's ability to refresh and brick sync once the
-		// access token expires (~1h). Keep the old one in that case.
-		if result.RefreshToken != "" {
-			cred.RefreshToken = result.RefreshToken
-		}
-		cred.TokenExpiry = result.Expiry.Format(time.RFC3339)
-		err = m.app.Accounts.StoreCredential(
-			context.Background(), p.accountID, p.cred.AccountFingerprint,
-			cred, credStore,
-		)
-		return storedMsg(err)
-	}
-}
-
-func (m Model) finishOAuthCredentialStore(msg oauthCredentialStoredMsg) (Model, tea.Cmd) {
-	m.oauthPending = false
-	if msg.err != nil {
-		m.statusToken++
-		m.syncStatus = fmt.Sprintf(
-			"Authorized, but storing the tokens failed: %s — re-authenticating again is safe",
-			msg.err,
-		)
-		m = m.reopenAccountSettings(msg.accountID)
-		return m, m.expireStatusAfter(10*time.Second, m.statusToken)
-	}
-	m = m.reopenAccountSettings(msg.accountID)
-	m.syncing = true
-	m.syncStatus = fmt.Sprintf("Syncing %s…", syncProgressLabel(msg.name))
-	return m, tea.Batch(
-		m.syncSpinner.Tick,
-		m.runSyncAccount(msg.accountID, msg.name),
-	)
-}
-
-// updateAccountCredentials rotates one account's secret in place. The stored
-// credential is loaded so its non-secret identity (username, client config) is
-// kept. Only the password (basic) or access token (bearer) is replaced.
-// StoreCredential re-checks the account fingerprint under the lifecycle lock.
-// A concurrent rename or removal then aborts the write instead of a corrupt
-// write.
-func (m Model) updateAccountCredentials(configured account.Account, secret string) tea.Cmd {
-	storedMsg := func(err error) accountCredentialStoredMsg {
-		return accountCredentialStoredMsg{
-			accountID: configured.ID,
-			name:      configured.DisplayName,
-			err:       err,
-		}
-	}
-	return func() tea.Msg {
-		ctx := context.Background()
-		credStore, err := m.openCredentialStore()
-		if err != nil {
-			return storedMsg(err)
-		}
-		fingerprint := configured.CredentialFingerprint()
-		cred, err := credentialForRotation(credStore.Get(configured.ID, fingerprint))
-		if err != nil {
-			return storedMsg(err)
-		}
-		if accountAuthIsBearer(configured.AuthType) {
-			cred.AccessToken = secret
-		} else {
-			cred.Password = secret
-		}
-		err = m.app.Accounts.StoreCredential(ctx, configured.ID, fingerprint, cred, credStore)
-		return storedMsg(err)
-	}
-}
-
-// credentialForRotation maps the pre-rotation Get outcome to the credential
-// the new secret is written into. A keyring entry that is gone or identity-
-// mismatched is one of the broken states rotation exists to repair. Those
-// start from an empty credential instead of a refusal. StoreCredential
-// re-seeds the account identity. Any other error aborts the rotation.
-func credentialForRotation(cred auth.Credential, err error) (auth.Credential, error) {
-	if err == nil {
-		return cred, nil
-	}
-	if auth.IsCredentialNotFound(err) || errors.Is(err, auth.ErrCredentialIdentityMismatch) {
-		return auth.Credential{}, nil
-	}
-	return auth.Credential{}, fmt.Errorf("load current credentials: %w", err)
-}
-
-// finishAccountCredentialStore lands the in-place rotation. Success reopens
-// Account Settings and syncs to confirm the new secret works; failure reopens
-// Settings with the error and leaves the previous credential intact.
-func (m Model) finishAccountCredentialStore(msg accountCredentialStoredMsg) (Model, tea.Cmd) {
-	if msg.err != nil {
-		m.syncing = false
-		m.statusToken++
-		m.syncStatus = fmt.Sprintf(
-			"Couldn't update credentials: %s — the previous sign-in is unchanged",
-			msg.err,
-		)
-		m = m.reopenAccountSettings(msg.accountID)
-		return m, m.expireStatusAfter(10*time.Second, m.statusToken)
-	}
-	m = m.reopenAccountSettings(msg.accountID)
-	m.syncing = true
-	m.syncStatus = fmt.Sprintf("Syncing %s…", syncProgressLabel(msg.name))
-	return m, tea.Batch(
-		m.syncSpinner.Tick,
-		m.runSyncAccount(msg.accountID, msg.name),
-	)
-}
-
-func calendarDiscoveryAccountName(accounts []account.Account, username string) string {
-	base := account.SuggestedName(username)
-	if base == "" {
-		base = "CalDAV"
-	}
-	taken := make(map[string]struct{}, len(accounts))
-	for _, configured := range accounts {
-		name := strings.TrimSpace(configured.DisplayName)
-		if name == "" {
-			name = account.UserFacingName(configured.Name, configured.Username, configured.ID)
-		}
-		taken[strings.ToLower(name)] = struct{}{}
-	}
-	if _, exists := taken[strings.ToLower(base)]; !exists {
-		return base
-	}
-	for suffix := 2; ; suffix++ {
-		candidate := fmt.Sprintf("%s (%d)", base, suffix)
-		if _, exists := taken[strings.ToLower(candidate)]; !exists {
-			return candidate
-		}
-	}
-}
-func existingCalendarDiscoveryAccount(accounts []account.Account, req CalendarDiscoveryRequestedMsg) (account.Account, bool) {
-	fingerprint := auth.AccountFingerprint(req.ServerURL, req.AuthType, req.Username)
-	for _, configured := range accounts {
-		if configured.CredentialFingerprint() == fingerprint {
-			return configured, true
-		}
-	}
-	return account.Account{}, false
-}
-
-func (m Model) connectAndDiscoverCalendar(req CalendarDiscoveryRequestedMsg, cred auth.Credential) tea.Cmd {
-	return func() tea.Msg {
-		ctx, cancel := context.WithTimeout(context.Background(), 2*time.Minute)
-		defer cancel()
-		store, err := m.openCredentialStore()
-		if err != nil {
-			return accountDiscoveryReadyMsg{err: err}
-		}
-		configured, err := m.app.Accounts.List(ctx)
-		if err != nil {
-			return accountDiscoveryReadyMsg{err: fmt.Errorf("list accounts: %w", err)}
-		}
-		if existing, ok := existingCalendarDiscoveryAccount(configured, req); ok {
-			discovery, err := m.app.Accounts.DiscoverWithCredential(ctx, existing.ID, cred, store)
-			if err != nil {
-				return accountDiscoveryReadyMsg{err: err}
-			}
-			return accountDiscoveryReadyMsg{discovery: discovery}
-		}
-		created, err := m.app.Accounts.Create(ctx, account.CreateParams{
-			Name:          calendarDiscoveryAccountName(configured, req.Username),
-			ServerURL:     req.ServerURL,
-			AuthType:      req.AuthType,
-			Username:      req.Username,
-			AllowInsecure: req.AllowInsecure,
-		}, cred, store)
-		if err != nil {
-			return accountDiscoveryReadyMsg{err: err}
-		}
-		discovery, err := m.app.Accounts.Discover(ctx, created.ID, store)
-		if err != nil {
-			if cleanupErr := m.app.Accounts.Delete(ctx, created.ID, store); cleanupErr != nil {
-				err = fmt.Errorf("%w (remove incomplete connection: %w)", err, cleanupErr)
-			}
-			return accountDiscoveryReadyMsg{err: err}
-		}
-		return accountDiscoveryReadyMsg{discovery: discovery, createdAccount: true}
-	}
-}
-
-func (m Model) finishOAuthCalendarDiscovery(result *auth.GoogleOAuthResult) tea.Cmd {
-	req := m.oauthPurpose.calendarDiscoveryMsg
-	cred := auth.Credential{
-		Username:          req.Username,
-		AccessToken:       result.AccessToken,
-		RefreshToken:      result.RefreshToken,
-		TokenExpiry:       result.Expiry.Format(time.RFC3339),
-		OAuthClientID:     req.OAuthClientID,
-		OAuthClientSecret: req.OAuthClientSecret,
-	}
-	return m.connectAndDiscoverCalendar(req, cred)
-}
-
-func (m Model) importAndSyncAccountCalendars(paths []string) tea.Cmd {
-	if m.calendarManager.DiscoveryPicker() == nil {
-		return func() tea.Msg {
-			return accountImportFinishedMsg{err: fmt.Errorf("calendar discovery is no longer open")}
-		}
-	}
-	discovery := m.calendarManager.DiscoveryPicker().discovery
-	return func() tea.Msg {
-		ctx, cancel := context.WithTimeout(context.Background(), 5*time.Minute)
-		defer cancel()
-		result, err := m.app.Accounts.Import(ctx, discovery, paths)
-		if err != nil {
-			return accountImportFinishedMsg{err: err}
-		}
-		finished := accountImportFinishedMsg{
-			created:  len(result.CreatedIDs),
-			existing: len(result.ExistingIDs),
-		}
-		if len(result.CreatedIDs) == 0 {
-			return finished
-		}
-		syncService, err := m.newSyncService()
-		if err != nil {
-			finished.syncErr = err
-			return finished
-		}
-		finished.synced, finished.warnings, finished.syncErr = syncNewlyLinkedCalendars(ctx, syncService, result.CreatedIDs, m.fullSyncStrategy)
-		return finished
-	}
-}
-
-func (m Model) reconcileAndSyncAccountCalendars(selection *accountCalendarSelection) tea.Cmd {
-	finished := accountSelectionFinishedMsg{
-		removedCurrent:    selection.removedCurrent,
-		accountManagement: selection.accountManagement,
-	}
-	discovery := selection.discovery
-	params := selection.params
-	return func() tea.Msg {
-		ctx, cancel := context.WithTimeout(context.Background(), 5*time.Minute)
-		defer cancel()
-		store, err := m.openCredentialStore()
-		if err != nil {
-			finished.err = err
-			return finished
-		}
-		result, err := m.app.Accounts.ReconcileSelection(ctx, discovery, params, store)
-		if err != nil {
-			finished.err = err
-			return finished
-		}
-		finished.created = len(result.CreatedIDs)
-		finished.removed = len(result.RemovedIDs)
-		finished.removedIDs = result.RemovedIDs
-		finished.accountRemoved = result.AccountRemoved
-		if len(result.CreatedIDs) == 0 {
-			return finished
-		}
-		syncService, err := m.newSyncService()
-		if err != nil {
-			finished.syncErr = err
-			return finished
-		}
-		finished.synced, finished.warnings, finished.syncErr = syncNewlyLinkedCalendars(ctx, syncService, result.CreatedIDs, m.fullSyncStrategy)
-		return finished
-	}
-}
-
-func (m Model) prepareAccountCalendarSelection(
-	msg AccountCalendarsReconcileRequestedMsg,
-) (*accountCalendarSelection, []accountDefaultCandidate, error) {
-	picker := m.calendarManager.DiscoveryPicker()
-	if picker == nil || !picker.manage || msg.AccountID != picker.discovery.Account.ID {
-		return nil, nil, fmt.Errorf("calendar selection no longer matches the open account")
-	}
-
-	discoveredByPath := make(map[string]account.DiscoveredCalendar, len(picker.discovery.Calendars))
-	for _, item := range picker.discovery.Calendars {
-		discoveredByPath[item.Path] = item
-	}
-	selected := make(map[string]struct{}, len(msg.SelectedPaths))
-	selectedPaths := make([]string, 0, len(msg.SelectedPaths))
-	for _, path := range msg.SelectedPaths {
-		if _, ok := discoveredByPath[path]; !ok {
-			return nil, nil, fmt.Errorf("calendar %q was not part of the open discovery", path)
-		}
-		if _, duplicate := selected[path]; duplicate {
-			continue
-		}
-		selected[path] = struct{}{}
-		selectedPaths = append(selectedPaths, path)
-	}
-
-	selection := &accountCalendarSelection{
-		discovery:         picker.discovery,
-		params:            account.SelectionParams{SelectedPaths: selectedPaths},
-		accountManagement: true,
-	}
-	removedIDs := make(map[int64]struct{})
-	var removedDefault bool
-	for _, item := range picker.discovery.Calendars {
-		if !item.Imported {
-			continue
-		}
-		if _, keep := selected[item.Path]; keep {
-			continue
-		}
-		selection.removed = append(selection.removed, item)
-		removedIDs[item.CalendarID] = struct{}{}
-		if m.calendarManagerOpen && m.calendarManager.LocalDraft() != nil &&
-			item.CalendarID == m.calendarManager.LocalDraft().ID {
-			selection.removedCurrent = true
-		}
-		if info, ok := m.calendars[item.CalendarID]; ok && info.IsDefault {
-			removedDefault = true
-		}
-	}
-	if !removedDefault {
-		return selection, nil, nil
-	}
-
-	candidates := make([]accountDefaultCandidate, 0, len(m.calendars)+len(picker.discovery.Calendars))
-	for id, info := range m.calendars {
-		if _, removed := removedIDs[id]; removed {
-			continue
-		}
-		candidates = append(candidates, accountDefaultCandidate{id: id, name: info.Name})
-	}
-	for _, item := range picker.discovery.Calendars {
-		if item.Imported {
-			continue
-		}
-		if _, add := selected[item.Path]; add {
-			candidates = append(candidates, accountDefaultCandidate{path: item.Path, name: item.Name})
-		}
-	}
-	slices.SortFunc(candidates, func(a, b accountDefaultCandidate) int {
-		if byName := strings.Compare(strings.ToLower(a.name), strings.ToLower(b.name)); byName != 0 {
-			return byName
-		}
-		if byPath := strings.Compare(a.path, b.path); byPath != 0 {
-			return byPath
-		}
-		switch {
-		case a.id < b.id:
-			return -1
-		case a.id > b.id:
-			return 1
-		default:
-			return 0
-		}
-	})
-	if len(candidates) == 0 {
-		return nil, nil, calendar.ErrLastCalendar
-	}
-	return selection, candidates, nil
-}
-
-func (m Model) showAccountCalendarRemovalConfirmation(selection *accountCalendarSelection) Model {
-	names := make([]string, 0, len(selection.removed))
-	for _, item := range selection.removed {
-		names = append(names, textsafe.Display(item.Name))
-	}
-	var message string
-	if len(names) == 1 {
-		message = fmt.Sprintf("Remove “%s” from Chroncal?", names[0])
-	} else {
-		message = fmt.Sprintf("Remove %d calendars from Chroncal?\n\n• %s",
-			len(names), strings.Join(names, "\n• "))
-	}
-	message += "\n\nNothing will be deleted from the server."
-	message += "\nLocal copies and changes not yet uploaded will be removed."
-	message += "\nYou can add these calendars again later."
-	buttonLabel := "Save Changes"
-	if len(selection.params.SelectedPaths) == 0 {
-		message += "\n\nThis also removes the account and stored sign-in from Chroncal."
-		buttonLabel = "Remove Account"
-	}
-	if selection.removedCurrent {
-		message += "\n\nYour unsaved calendar edits will be discarded."
-	}
-	m.pendingAccountSelection = selection
-	m.confirmDialog = NewConfirmDialogModel(message, buttonLabel, m.theme).
-		Destructive().
-		SetSize(m.width, m.height)
-	m.confirmOpen = true
-	return m
-}
-func (m Model) discardDiscoveryAccount(accountID int64) tea.Cmd {
-	return func() tea.Msg {
-		ctx, cancel := context.WithTimeout(context.Background(), 2*time.Minute)
-		defer cancel()
-		store, err := m.openCredentialStore()
-		if err == nil {
-			err = m.app.Accounts.Delete(ctx, accountID, store)
-		}
-		return calendarDiscoveryDiscardedMsg{err: err}
-	}
-}
-
-func (m Model) removeAccount(accountID int64, name string) tea.Cmd {
-	return func() tea.Msg {
-		ctx, cancel := context.WithTimeout(context.Background(), 2*time.Minute)
-		defer cancel()
-		store, err := m.openCredentialStore()
-		if err == nil {
-			err = m.app.Accounts.Delete(ctx, accountID, store)
-		}
-		return accountRemovalFinishedMsg{accountID: accountID, name: name, err: err}
-	}
-}
-
-func (m Model) discoverAccountCalendars(accountID int64, generation uint64) tea.Cmd {
-	result := func(discovery account.Discovery, err error) accountManagementDiscoveryReadyMsg {
-		return accountManagementDiscoveryReadyMsg{
-			discovery: discovery, accountID: accountID, generation: generation, err: err,
-		}
-	}
-	return func() tea.Msg {
-		ctx, cancel := context.WithTimeout(context.Background(), 2*time.Minute)
-		defer cancel()
-		store, err := m.openCredentialStore()
-		if err != nil {
-			return result(account.Discovery{}, fmt.Errorf("open credential store: %w", err))
-		}
-		discovery, err := m.app.Accounts.Discover(ctx, accountID, store)
-		return result(discovery, err)
-	}
-}
-
-// syncProgressLabel produces a short calendar label for the per-calendar
-// progress footer. Calendar names can be long (Apple/Google often exceed 40
-// chars) and would push the spinner off the visible status line.
-func syncProgressLabel(name string) string {
-	if name == "" {
-		return "calendar"
-	}
-	const maxLen = 32
-	runes := []rune(name)
-	if len(runes) <= maxLen {
-		return name
-	}
-	return string(runes[:maxLen-1]) + "…"
-}
-
-// runOpportunisticPush pushes unpushed changes for a single calendar with no
-// pull. Best-effort: failures do not surface as errors. The dirty flag
-// survives and the background tick will retry. A 412 records a conflict and
-// keeps the saved edit; the footer then names the conflict count. Returns
-// nil for local-only calendars (Synced=false). Callers can then batch it
-// into the post-save command with no UI noise for offline calendars.
-func (m Model) runOpportunisticPush(calendarID int64) tea.Cmd {
-	info, ok := m.calendars[calendarID]
-	if !ok || !info.Synced {
-		return nil
-	}
-	name := info.Name
-	return func() tea.Msg {
-		svc, err := m.newSyncService()
-		if err != nil {
-			return opportunisticPushFinishedMsg{err: err}
-		}
-		ctx, cancel := context.WithTimeout(context.Background(), 30*time.Second)
-		defer cancel()
-		result, err := svc.PushLocalEdits(ctx, calendarID)
-		if err != nil {
-			return opportunisticPushFinishedMsg{err: err}
-		}
-		if result.Pushed == 0 && result.Deleted == 0 && len(result.Errors) == 0 && len(result.Warnings) == 0 &&
-			result.Conflicts == 0 && result.AutoResolved == 0 &&
-			result.SkippedConflicts == 0 {
-			return opportunisticPushFinishedMsg{}
-		}
-		label := name
-		if label == "" {
-			label = "calendar"
-		}
-		summary := syncSummary(label, syncTotals{
-			pushed:           result.Pushed,
-			deleted:          result.Deleted,
-			conflicts:        result.Conflicts,
-			autoResolved:     result.AutoResolved,
-			skippedConflicts: result.SkippedConflicts,
-			warnings:         len(result.Warnings),
-		})
-		var firstErr error
-		if len(result.Errors) > 0 {
-			firstErr = result.Errors[0]
-		}
-		return opportunisticPushFinishedMsg{summary: summary, err: firstErr}
-	}
-}
-
-// syncSummary builds the footer confirmation for a completed sync. It lists
-// only the non-zero counters. A no-op sync reads "Synced Work · up to date"
-// instead of five "· 0" segments behind it. t.warnings counts the
-// import warnings on the SyncResult: values the importer had to fabricate.
-// The count is the in-UI signal. The full text is in the state-dir log.
-// Take syncTotals rather than five positional ints. A transposed
-// counter then cannot compile in silence.
-func syncSummary(label string, t syncTotals) string {
-	var parts []string
-	if t.pushed > 0 {
-		parts = append(parts, fmt.Sprintf("pushed %d", t.pushed))
-	}
-	if t.pulled > 0 {
-		parts = append(parts, fmt.Sprintf("pulled %d", t.pulled))
-	}
-	if t.deleted > 0 {
-		parts = append(parts, fmt.Sprintf("deleted %d", t.deleted))
-	}
-	if t.conflicts > 0 {
-		parts = append(parts, fmt.Sprintf("%d %s", t.conflicts, pluralize(t.conflicts, "conflict", "conflicts")))
-	}
-	if t.autoResolved > 0 {
-		parts = append(parts, fmt.Sprintf("%d auto-resolved", t.autoResolved))
-	}
-	if t.skippedConflicts > 0 {
-		parts = append(parts, fmt.Sprintf("%d held by conflicts", t.skippedConflicts))
-	}
-	if t.warnings > 0 {
-		parts = append(parts, importWarningsSegment(t.warnings))
-	}
-	if len(parts) == 0 {
-		return fmt.Sprintf("Synced %s · up to date", label)
-	}
-	return fmt.Sprintf("Synced %s · %s", label, strings.Join(parts, " · "))
-}
-
-// syncNewlyLinkedCalendars runs the first pull for each calendar an account
-// link just created. It returns how many synced cleanly, the total import-
-// warning count (fabricated values the status line must mention), and the
-// first error. Later calendars still sync. One failure then does not strand
-// the rest unsynced.
-func syncNewlyLinkedCalendars(ctx context.Context, svc *syncpkg.Service, ids []int64, strategy syncpkg.ConflictStrategy) (synced, warnings int, syncErr error) {
-	for _, calendarID := range ids {
-		syncResult, err := svc.SyncCalendar(ctx, calendarID, strategy)
-		if syncResult != nil {
-			warnings += len(syncResult.Warnings)
-		}
-		if err == nil && len(syncResult.Errors) > 0 {
-			err = errors.Join(syncResult.Errors...)
-		}
-		if err != nil {
-			if syncErr == nil {
-				syncErr = err
-			}
-			continue
-		}
-		synced++
-	}
-	return synced, warnings, syncErr
-}
-
-// importWarningsSegment renders the "N import warning(s)" counter shared by
-// syncSummary and the account-link status lines.
-func importWarningsSegment(warnings int) string {
-	return fmt.Sprintf("%d %s", warnings, pluralize(warnings, "import warning", "import warnings"))
-}
-
-// appendImportWarnings suffixes a status line with the import-warning count
-// when there is one. Account-link pulls then surface fabricated values the
-// same way syncSummary does for every other TUI sync path.
-func appendImportWarnings(status string, warnings int) string {
-	if warnings <= 0 {
-		return status
-	}
-	return status + " · " + importWarningsSegment(warnings)
-}
-
-func (m Model) expireStatusAfter(d time.Duration, token int) tea.Cmd {
-	return tea.Tick(d, func(time.Time) tea.Msg {
-		return syncStatusExpiredMsg{token: token}
-	})
-}
-
-func nextClockTickDelay(now time.Time) time.Duration {
-	nextMinute := now.Truncate(time.Minute).Add(time.Minute)
-	d := nextMinute.Sub(now)
-	if d <= 0 {
-		return time.Minute
-	}
-	return d
-}
-
-func nextClockTick(token int) tea.Cmd {
-	return tea.Tick(nextClockTickDelay(time.Now()), func(time.Time) tea.Msg {
-		return clockTickMsg{token: token}
-	})
-}
-
-// scheduleClockTick arms the once-a-minute tick if one is not already armed.
-// Every view runs it. Day and week need it to advance the now-line. All views
-// need it so the "today" cell highlight follows the midnight day rollover.
-func (m Model) scheduleClockTick() (Model, tea.Cmd) {
-	if m.clockTickScheduled {
-		return m, nil
-	}
-	m.clockTickToken++
-	m.clockTickScheduled = true
-	return m, nextClockTick(m.clockTickToken)
-}
-
-// refreshToday recomputes the local "today" date and propagates it into every
-// view model. The views render their "today" highlight from these stored
-// fields. Without this they freeze on the date captured at startup once the
-// app is left open across midnight. Day view's grid and the sidebar mini-month
-// recompute today at render time and are already immune. A refresh of the
-// stored fields keeps week, month, and agenda current. A later view switch
-// then carries the right date.
-func (m Model) refreshToday(now time.Time) Model {
-	today := time.Date(now.Year(), now.Month(), now.Day(), 0, 0, 0, 0, now.Location())
-	m.day.today = today
-	m.week.today = today
-	m.agenda.today = today
-	m.calendar.today = today
-	return m
-}
-
-// navigateMainTo sets the active main view's cursor (and the month-view's
-// displayed month) to the given date. Callers typically follow this with
-// m.loadEvents() to refresh the query range.
-func (m Model) navigateMainTo(t time.Time) Model {
-	switch m.viewMode {
-	case viewDay:
-		m.day.cursor = t
-	case viewWeek:
-		m.week.cursor = t
-	case viewAgenda:
-		m.agenda.cursor = t
-		m.agenda = m.agenda.ResetWindow(t)
-	default:
-		m.calendar.cursor = t
-		m.calendar.month = time.Date(t.Year(), t.Month(), 1, 0, 0, 0, 0, t.Location())
-	}
-	return m
-}
-
-// refreshCalendarViews recomputes the per-view CalendarEvent slices from the
-// current m.events using the current m.hiddenCalendars set. Use this after the
-// hidden set changes (no DB round-trip needed).
-func (m Model) refreshCalendarViews() Model {
-	calEvents := eventsToCalendar(m.events, m.calendars, m.hiddenCalendars)
-	m.calendar = m.calendar.SetEvents(calEvents)
-	m.week = m.week.SetEvents(calEvents)
-	m.day = m.day.SetEvents(calEvents)
-	m.agenda = m.agenda.SetEvents(filterVisibleEvents(m.events, m.hiddenCalendars), m.calendars)
-	return m
-}
-
-// filterVisibleEvents drops events whose calendar is currently hidden. The
-// agenda row list then stays in sync with the toggle set. The original slice
-// is not mutated.
-func filterVisibleEvents(events []event.Event, hidden map[int64]bool) []event.Event {
-	if len(hidden) == 0 {
-		return events
-	}
-	out := make([]event.Event, 0, len(events))
-	for _, e := range events {
-		if hidden[e.CalendarID] {
-			continue
-		}
-		out = append(out, e)
-	}
-	return out
-}
-
 func (m Model) Init() tea.Cmd {
 	cmds := []tea.Cmd{tea.RequestBackgroundColor, m.loadEvents(), m.loadCalendars()}
 	if m.clockTickScheduled {
 		cmds = append(cmds, nextClockTick(m.clockTickToken))
 	}
 	return tea.Batch(cmds...)
-}
-
-const sidebarWidth = 24
-
-// footerHeight returns the total rows the footer occupies. The footer is
-// always a single line: the "? help" hint (and optional sync status).
-func (m Model) footerHeight() int {
-	return 1
-}
-
-func (m Model) mainDims() (int, int) {
-	padding := 1
-	contentHeight := m.height - m.footerHeight()
-	mainWidth := m.width - padding*2
-	if m.showSidebar {
-		mainWidth -= sidebarWidth
-	}
-	return mainWidth, contentHeight
-}
-
-func (m Model) calendarOffset() (int, int) {
-	padding := 1
-	x := padding
-	if m.showSidebar {
-		x += sidebarWidth
-	}
-	return x, padding
-}
-
-// innerDims returns the space available inside the padded main box,
-// which is what the calendar renderer should fill.
-func (m Model) innerDims() (int, int) {
-	mw, mh := m.mainDims()
-	padding := 1
-	return mw - padding*2, mh - padding*2
-}
-
-// clearConfirmPending drops wait state owned by a destructive confirm or
-// choice. Used when ctrl+c abandons that operation in favor of the quit
-// confirm. The superseded action can then never fire or reappear afterward.
-func (m Model) clearConfirmPending() Model {
-	m.pendingDelete = event.Event{}
-	m.pendingPurgeEntries = nil
-	m.pendingPurgeTitle = ""
-	m.pendingCalendarDelete = 0
-	m.pendingCalendarDeleteName = ""
-	m.pendingCalendarPromote = 0
-	m.pendingCalendarPromoteName = ""
-	m.pendingCalendarPromoteCands = nil
-	m.pendingAccountSelection = nil
-	m.pendingAccountDefaultCandidates = nil
-	m.pendingAccountRemoveID = 0
-	m.pendingAccountRemoveName = ""
-	if m.choiceOpen {
-		m.pendingCalendarMove = nil
-	}
-	m.choiceOpen = false
-	m.pendingScopeKind = pendingScopeNone
-	return m
-}
-
-// openQuitConfirm builds and opens the quit-confirm dialog. Shared by the
-// q and ctrl+c entry points so the two keystrokes cannot drift in style.
-func (m Model) openQuitConfirm() Model {
-	m.pendingQuit = true
-	m.confirmDialog = NewConfirmDialogModel("Quit chroncal?", "Quit", m.theme).
-		SetSize(m.width, m.height)
-	m.confirmOpen = true
-	return m
-}
-
-// interceptGlobalKeys routes the quit guard (q / ctrl+c) and help (?) ahead
-// of any open dialog so they work from anywhere. A second ctrl+c while the
-// quit confirm is on screen forces the exit. ctrl+c is truly global. It is
-// not a character anyone types into a field. ? is suppressed while a
-// text-entry surface owns input (palette search, event form, calendar form)
-// so users can type it normally. q is suppressed while any overlay is open
-// so the overlay's own close binding runs instead. The quit confirm also
-// blocks ?. The help dialog handles its own close keys.
-func (m Model) interceptGlobalKeys(msg tea.KeyPressMsg) (Model, tea.Cmd, bool) {
-	inQuitConfirm := m.confirmOpen && m.pendingQuit
-	if msg.String() == "ctrl+c" {
-		if inQuitConfirm {
-			m.oauthFlow.Abort() // release any in-flight OAuth listener
-			return m, tea.Quit, true
-		}
-		// ctrl+c is truly global. Even when a destructive (non-quit)
-		// confirm — event delete, trash purge, calendar delete — owns
-		// input, replace it with the quit confirm. Do not let the
-		// keystroke fall through to confirmDialog.Update (which ignores
-		// it). Clear the abandoned confirm's pending state. That keeps the
-		// destructive action from a later fire.
-		m = m.clearConfirmPending()
-		return m.openQuitConfirm(), nil, true
-	}
-	textEntryActive := m.paletteOpen || m.formOpen || m.calendarManagerOpen ||
-		m.oauthFlowOpen || m.accountOAuthConfigOpen || m.accountCredentialsOpen
-	// q opens the quit confirm only from the bare grid. Any open overlay
-	// owns q. That includes text-entry (palette, form) and read-only/choice
-	// (event view, list, choice, calendar list, trash, help). Its own
-	// `q`-to-close key binding then runs, not the global quit guard
-	// (issue #406).
-	if key.Matches(msg, m.keys.Quit) && !m.anyOverlayOpen() {
-		return m.openQuitConfirm(), nil, true
-	}
-	if key.Matches(msg, m.keys.Help) && !inQuitConfirm && !m.helpDialogOpen && !textEntryActive {
-		return m, func() tea.Msg { return HelpDialogRequestedMsg{} }, true
-	}
-	// Trash: shift+D opens the Recently-deleted overlay from the main grid.
-	// Blocked while a text-entry surface owns input (so typing "D" in the
-	// palette / form doesn't jump out).
-	if key.Matches(msg, m.keys.TrashView) && !m.trashOpen && !inQuitConfirm && !textEntryActive && !m.anyOverlayOpen() {
-		m.trashOpen = true
-		m.trash = NewTrashModel(m.calendars, newThemedHelp(m.theme)).
-			SetSelectedColor(m.theme.Selected).
-			SetSize(m.width, m.height)
-		return m, m.loadTrash(), true
-	}
-	// Undo: only active on the main grid, with no overlay competing for input.
-	if key.Matches(msg, m.keys.Undo) && m.undoIsAllowed() {
-		entry, ok := m.undoStack.Peek()
-		if ok {
-			// Bumping the token invalidates any delete-push that was still
-			// waiting for the 6-second window to elapse.
-			m.pushDeferralToken++
-			m.undoRestoreInFlight = true
-			m.toast.Restoring()
-			meta := entry.Meta
-			title := meta.Label
-			cmd := func() tea.Msg {
-				err := m.app.Events.RestoreUndo(context.Background(), meta)
-				return eventRestoredMsg{meta: meta, title: title, err: err}
-			}
-			return m, cmd, true
-		}
-	}
-	return m, nil, false
-}
-
-// currentFooterContext maps the app's focus/view state to a FooterContext,
-// the input the pure-render FooterModel wants.
-func (m Model) currentFooterContext() FooterContext {
-	switch {
-	case m.calendarManagerOpen:
-		return FooterCalendarPopup
-	case m.viewDialogOpen:
-		return FooterEventPopup
-	case m.focus == focusSidebar:
-		return FooterSidebar
-	}
-	switch m.viewMode {
-	case viewAgenda:
-		if _, ok := m.agenda.SelectedEvent(); ok {
-			return FooterAgenda
-		}
-		return FooterAgendaEmpty
-	case viewWeek:
-		return FooterWeek
-	case viewDay:
-		return FooterDay
-	default:
-		return FooterMonth
-	}
-}
-
-// currentFooterHasRSVP reports whether the event-popup footer should advertise
-// RSVP keys. Only meaningful when the event view dialog is open.
-func (m Model) currentFooterHasRSVP() bool {
-	if !m.viewDialogOpen {
-		return false
-	}
-	// The event view dialog exposes RSVP only when the user is an invited
-	// attendee; defer to its own rsvpActions helper via the dialog model.
-	return len(m.viewDialog.rsvpActions()) > 0
-}
-
-// currentFooterShowsTodayHint reports whether the active view's selected day
-// differs from today, making the `t today` footer hint actionable.
-func (m Model) currentFooterShowsTodayHint() bool {
-	switch m.currentFooterContext() {
-	case FooterMonth, FooterWeek, FooterDay, FooterAgendaEmpty:
-		cursor, today := m.viewCursorAndToday()
-		return !sameDay(cursor, today)
-	case FooterAgenda:
-		// Agenda navigation moves the selected row, not m.cursor. Look
-		// at the selected day. If nothing is selected, use the cursor.
-		// Then compare against today.
-		_, today := m.viewCursorAndToday()
-		return !sameDay(m.agenda.SelectedDay(), today)
-	default:
-		return false
-	}
-}
-
-// undoIsAllowed reports whether the `u` key should trigger an undo. The guard
-// is intentionally strict. Any overlay, editor, or palette that might consume
-// character input takes priority. Otherwise a stray `u` in a title field
-// would trigger a restore in silence.
-func (m Model) undoIsAllowed() bool {
-	if m.focus != focusCalendar {
-		return false
-	}
-	if m.anyOverlayOpen() {
-		return false
-	}
-	if m.undoRestoreInFlight {
-		return false
-	}
-	return m.undoStack != nil && m.undoStack.Len() > 0
-}
-
-// anyOverlayOpen reports whether any dialog, form, or palette is currently
-// on screen. While one is up it owns input and renders its own help row.
-// The app footer should then degrade to status plus toast rather than a
-// duplicate of the dialog's hints.
-func (m Model) anyOverlayOpen() bool {
-	return m.paletteOpen || m.formOpen || m.viewDialogOpen || m.dialogOpen ||
-		m.confirmOpen || m.choiceOpen || m.calendarManagerOpen ||
-		m.accountRenameOpen || m.accountOAuthConfigOpen || m.accountCredentialsOpen ||
-		m.helpDialogOpen || m.trashOpen || m.oauthFlowOpen
-}
-
-func (m Model) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
-	if msg, ok := msg.(clockTickMsg); ok {
-		if msg.token != m.clockTickToken {
-			return m, nil
-		}
-		m.clockTickScheduled = false
-		m = m.refreshToday(time.Now())
-		return m.scheduleClockTick()
-	}
-
-	switch msg := msg.(type) {
-	case syncFinishedMsg:
-		return m.finishSync(msg)
-	case oauthCredentialStoredMsg:
-		return m.finishOAuthCredentialStore(msg)
-	case accountCredentialStoredMsg:
-		return m.finishAccountCredentialStore(msg)
-	case accountReauthReadyMsg:
-		m.oauthPending = false
-		configured, ok := m.accounts[msg.accountID]
-		if msg.err == nil && (!ok || configured.ID == 0 || !calendarAuthIsOAuth(configured.AuthType)) {
-			msg.err = fmt.Errorf("account is no longer available for OAuth")
-		}
-		if msg.err != nil {
-			if params, ok := m.accountSettingsParams(msg.accountID); ok {
-				m.calendarManager = m.calendarManager.OpenAccount(params).SetSize(m.width, m.height)
-				m.calendarManagerOpen = true
-			}
-			m.statusToken++
-			m.syncStatus = "Re-authentication failed: " + msg.err.Error()
-			return m, m.expireStatusAfter(8*time.Second, m.statusToken)
-		}
-		if msg.cred.OAuthClientID == "" || msg.cred.OAuthClientSecret == "" {
-			m.accountOAuthConfig = NewAccountOAuthConfigDialogModel(
-				msg.accountID, msg.name, msg.cred.OAuthClientID, m.theme,
-			).SetSize(m.width, m.height)
-			m.accountOAuthConfigOpen = true
-			m.syncStatus = ""
-			return m, nil
-		}
-		m.oauthPurpose = oauthFlowPurpose{
-			accountID: msg.accountID, accountName: msg.name, cred: msg.cred,
-		}
-		return m.startOAuthFlow(msg.cred.OAuthClientID, msg.cred.OAuthClientSecret)
-	case accountManagementDiscoveryReadyMsg:
-		return m.finishAccountManagementDiscovery(msg)
-	case calendarMoveDiscoveryReadyMsg:
-		return m.finishCalendarMoveDiscovery(msg)
-	case calendarMoveFinishedMsg:
-		return m.finishCalendarMove(msg)
-	case accountRenameFinishedMsg:
-		return m.finishAccountRename(msg)
-	case accountRemovalFinishedMsg:
-		return m.finishAccountRemoval(msg)
-	}
-
-	// Global key bindings override any open dialog: ctrl+c / q always route
-	// through the quit guard, and ? opens the help dialog. The quit confirm
-	// itself is exempt so its y/n/esc keys keep working, and ? is a no-op
-	// while the help dialog is already up (it handles its own close keys).
-	if kp, ok := msg.(tea.KeyPressMsg); ok {
-		if newM, cmd, handled := m.interceptGlobalKeys(kp); handled {
-			return newM, cmd
-		}
-	}
-
-	// The quit guard sits on top of every other overlay (including help,
-	// palette, and text-entry dialogs), so it must own input whenever it's
-	// up. Without this, keystrokes like y/n/esc would be swallowed by
-	// whatever dialog happens to be underneath.
-	if m.confirmOpen && m.pendingQuit {
-		switch msg.(type) {
-		case tea.KeyPressMsg, tea.MouseClickMsg, tea.MouseWheelMsg, tea.MouseMotionMsg, tea.MouseReleaseMsg, tea.PasteMsg:
-			var cmd tea.Cmd
-			m.confirmDialog, cmd = m.confirmDialog.Update(msg)
-			return m, cmd
-		}
-	}
-
-	// The OAuth pending modal owns key and mouse input while it's up
-	// (below only the quit guard). Flow lifecycle messages, loads, and
-	// sizes fall through to the main switch, which routes them to the
-	// modal and dispatches the outcome.
-	if m.oauthFlowOpen && (!m.confirmOpen || !m.pendingQuit) {
-		switch msg.(type) {
-		case tea.KeyPressMsg, tea.MouseClickMsg:
-			var cmd tea.Cmd
-			m.oauthFlow, cmd = m.oauthFlow.Update(msg)
-			return m, cmd
-		}
-	}
-
-	// When the palette is open, it captures all input. Only specific
-	// parent-level messages (size, theme, palette-result, and
-	// spinner ticks) fall through.
-	if m.paletteOpen {
-		switch msg.(type) {
-		case PaletteSelectedMsg, PaletteClosedMsg,
-			tea.BackgroundColorMsg, tea.WindowSizeMsg,
-			spinner.TickMsg:
-			// fall through to main switch
-		default:
-			var cmd tea.Cmd
-			m.palette, cmd = m.palette.Update(msg)
-			return m, cmd
-		}
-	}
-
-	if m.accountOAuthConfigOpen {
-		switch msg.(type) {
-		case AccountOAuthConfigSubmittedMsg, AccountOAuthConfigClosedMsg,
-			syncStatusExpiredMsg, toastTickMsg, spinner.TickMsg,
-			tea.BackgroundColorMsg, tea.WindowSizeMsg:
-			// fall through to main switch
-		default:
-			var cmd tea.Cmd
-			m.accountOAuthConfig, cmd = m.accountOAuthConfig.Update(msg)
-			return m, cmd
-		}
-	}
-
-	if m.accountCredentialsOpen {
-		switch msg.(type) {
-		case AccountCredentialsUpdateSubmittedMsg, AccountCredentialsUpdateClosedMsg,
-			syncStatusExpiredMsg, toastTickMsg, spinner.TickMsg,
-			tea.BackgroundColorMsg, tea.WindowSizeMsg:
-			// fall through to main switch
-		default:
-			var cmd tea.Cmd
-			m.accountCredentials, cmd = m.accountCredentials.Update(msg)
-			return m, cmd
-		}
-	}
-
-	if m.accountRenameOpen {
-		switch msg.(type) {
-		case AccountRenameRequestedMsg, accountRenameCancelledMsg,
-			accountRenameFinishedMsg, syncStatusExpiredMsg, toastTickMsg, spinner.TickMsg,
-			tea.BackgroundColorMsg, tea.WindowSizeMsg:
-			// fall through to main switch
-		default:
-			var cmd tea.Cmd
-			m.accountRename, cmd = m.accountRename.Update(msg)
-			return m, cmd
-		}
-	}
-
-	// Calendar manager is the canonical calendar management overlay.
-	// It hosts calendar detail, account detail, and discovery flows.
-	if m.calendarManagerOpen && !m.accountRenameOpen &&
-		!m.accountOAuthConfigOpen && !m.accountCredentialsOpen &&
-		!m.confirmOpen && !m.choiceOpen {
-		switch msg.(type) {
-		case CalendarManagerClosedMsg, CalendarManagerRequestedMsg, CalendarTransferClosedMsg,
-			CalendarSavedMsg, CalendarDiscoveryRequestedMsg,
-			CalendarDeleteRequestedMsg, CalendarKeepLocalRequestedMsg, CalendarMoveToAccountRequestedMsg, CalendarTestRequestedMsg,
-			CalendarVisibilityToggledMsg, CalendarReorderedMsg, AccountReorderedMsg,
-			calendarOrderSavedMsg, accountOrderSavedMsg,
-			CalendarExportRequestedMsg, CalendarImportPreviewRequestedMsg,
-			CalendarImportRequestedMsg, CalendarExportWriteRequestedMsg,
-			calendarImportPreviewReadyMsg, calendarImportFinishedMsg, calendarExportFinishedMsg,
-			AccountSettingsRequestedMsg, AccountSettingsClosedMsg, AccountSettingsManageRequestedMsg, AccountSettingsSyncRequestedMsg,
-			AccountSettingsRenameRequestedMsg, AccountSettingsReauthRequestedMsg,
-			AccountSettingsUpdateCredentialsRequestedMsg,
-			AccountSettingsRemoveRequestedMsg, AccountRenameRequestedMsg,
-			CalendarSetDefaultRequestedMsg,
-			AccountCalendarsImportRequestedMsg, AccountCalendarsReconcileRequestedMsg,
-			AccountCalendarPickerClosedMsg,
-			accountDiscoveryReadyMsg, accountImportFinishedMsg, accountSelectionFinishedMsg,
-			calendarDiscoveryDiscardedMsg,
-			calendarDeleteCountMsg,
-			tea.BackgroundColorMsg, tea.WindowSizeMsg,
-			eventsLoadedMsg, calendarsLoadedMsg,
-			calendarMutationDoneMsg,
-			toastTickMsg, spinner.TickMsg, syncStatusExpiredMsg:
-			// fall through to main switch
-		default:
-			var cmd tea.Cmd
-			m.calendarManager, cmd = m.calendarManager.Update(msg)
-			return m, cmd
-		}
-	}
-
-	// When the form is open, route most messages to it (cursor blink, keys, etc.).
-	// Only let specific parent-level messages fall through to the main switch.
-	if m.formOpen {
-		switch msg.(type) {
-		case EventFormSaveMsg, EventFormClosedMsg,
-			tea.BackgroundColorMsg, tea.WindowSizeMsg,
-			eventsLoadedMsg, calendarsLoadedMsg,
-			eventCreatedMsg, eventUpdatedMsg,
-			spinner.TickMsg:
-			// fall through to main switch
-		default:
-			var cmd tea.Cmd
-			m.form, cmd = m.form.Update(msg)
-			return m, cmd
-		}
-	}
-
-	switch msg := msg.(type) {
-	case tea.BackgroundColorMsg:
-		m.theme = LoadTheme(m.themeName, msg.IsDark())
-		// System theme opts into a terminal-adaptive Selected color.
-		// Derive it from the actual terminal background reported over
-		// OSC 11 by a shift of OKLCh lightness ±8 %. That guarantees a
-		// visible-but-subtle highlight on any terminal theme. The
-		// static Dracula Selection hexes in system.toml are the
-		// fallback when OSC 11 does not answer (e.g. tmux without
-		// passthrough).
-		if m.themeName == "system" && msg.Color != nil {
-			delta := 0.08
-			if !msg.IsDark() {
-				delta = -0.08
-			}
-			m.theme.Selected = oklch.ShiftLightness(msg.Color, delta)
-			// Secondary buttons need more weight than a list-row fill —
-			// at ±0.08 the pill barely separates from the terminal bg on
-			// low-contrast themes. Use ±0.18 so the button reads as a
-			// tappable surface, while list-row Selected stays subtle.
-			btnDelta := 0.18
-			if !msg.IsDark() {
-				btnDelta = -0.18
-			}
-			m.theme.ButtonBg = oklch.ShiftLightness(msg.Color, btnDelta)
-		}
-		SetActiveTheme(m.theme)
-		// Month/week/day views use the selected-color as a vibrant
-		// BORDER stroke around the cursor cell; Primary (the brand
-		// accent) always stands out against the cell background. The
-		// neutral theme.Selected is reserved for list-row FILLS
-		// (trash, event list, palette, …) where a muted highlight is
-		// what you want.
-		m.calendar = m.calendar.SetSelectedColor(m.theme.Primary)
-		m.week = m.week.SetSelectedColor(m.theme.Primary)
-		m.day = m.day.SetSelectedColor(m.theme.Primary)
-		m.agenda = m.agenda.SetTheme(m.theme)
-		m.sidebar = m.sidebar.SetTheme(m.theme)
-		m.toast.SetTheme(m.theme)
-		m.oauthFlow = m.oauthFlow.SetTheme(m.theme)
-		m.footer.SetTheme(m.theme)
-		m.calendarManager = m.calendarManager.SetTheme(m.theme)
-		if m.trashOpen {
-			m.trash = m.trash.SetSelectedColor(m.theme.Selected)
-		}
-		return m, nil
-
-	case tea.WindowSizeMsg:
-		m.width = msg.Width
-		m.height = msg.Height
-		iw, ih := m.innerDims()
-		m.calendar = m.calendar.SetSize(iw, ih)
-		m.week = m.week.SetSize(iw, ih)
-		m.day = m.day.SetSize(iw, ih)
-		m.agenda = m.agenda.SetSize(iw, ih)
-		m.trash = m.trash.SetSize(m.width, m.height)
-		m.dialog = m.dialog.SetSize(m.width, m.height)
-		m.viewDialog = m.viewDialog.SetSize(m.width, m.height)
-		m.confirmDialog = m.confirmDialog.SetSize(m.width, m.height)
-		m.choiceDialog = m.choiceDialog.SetSize(m.width, m.height)
-		m.calendarManager = m.calendarManager.SetSize(m.width, m.height)
-		m.accountRename = m.accountRename.SetSize(m.width, m.height)
-		m.accountOAuthConfig = m.accountOAuthConfig.SetSize(m.width, m.height)
-		m.accountCredentials = m.accountCredentials.SetSize(m.width, m.height)
-		m.form = m.form.SetSize(m.width, m.height)
-		m.palette = m.palette.SetSize(m.width, m.height)
-		m.helpDialog = m.helpDialog.SetSize(m.width, m.height)
-		m.oauthFlow = m.oauthFlow.SetSize(m.width, m.height)
-		m.ready = true
-		return m, nil
-
-	case eventsLoadedMsg:
-		// Guard against a stale load. Rapid navigation (e.g. repeated [/] in
-		// the agenda) fires multiple in-flight queries. Drop any whose range
-		// no longer matches the active view. A late stale response then cannot
-		// overwrite correct rows with an empty set.
-		//
-		// For full (merge=false) responses the query range must equal the
-		// current expected range exactly. For incremental (merge=true)
-		// responses the queried slice must lie inside the expected range
-		// and abut the currently-loaded range. The append is then meaningful.
-		// Otherwise the user has since jumped elsewhere.
-		expectedFrom, expectedTo := m.expectedEventRange()
-		if msg.merge {
-			if msg.from.Before(expectedFrom) || msg.to.After(expectedTo) {
-				return m, nil
-			}
-			if !msg.from.Equal(m.loadedTo) && !msg.to.Equal(m.loadedFrom) {
-				return m, nil
-			}
-		} else if !msg.from.Equal(expectedFrom) || !msg.to.Equal(expectedTo) {
-			return m, nil
-		}
-		m.err = msg.err
-		if msg.merge {
-			m.events = mergeEvents(m.events, msg.events)
-			if msg.from.Before(m.loadedFrom) {
-				m.loadedFrom = msg.from
-			}
-			if msg.to.After(m.loadedTo) {
-				m.loadedTo = msg.to
-			}
-		} else {
-			m.events = msg.events
-			m.loadedFrom = msg.from
-			m.loadedTo = msg.to
-		}
-		calEvents := eventsToCalendar(msg.events, m.calendars, m.hiddenCalendars)
-		switch m.viewMode {
-		case viewDay:
-			m.day = m.day.SetEvents(calEvents)
-		case viewWeek:
-			m.week = m.week.SetEvents(calEvents)
-		case viewAgenda:
-			// Pass m.events (the merged cache), not msg.events. msg.events
-			// is only the delta for incremental responses. Otherwise a
-			// merge would rebuild the agenda rows with only the new
-			// slice. The events shown before would vanish.
-			m.agenda = m.agenda.SetEvents(filterVisibleEvents(m.events, m.hiddenCalendars), m.calendars)
-		default:
-			m.calendar = m.calendar.SetEvents(calEvents)
-		}
-		if m.dialogOpen {
-			dayEvents := eventsOn(filterVisibleEvents(m.events, m.hiddenCalendars), m.dialog.day)
-			m.dialog = m.dialog.SetEvents(dayEvents)
-		}
-		// After an agenda load lands, pull in the next month if the loaded
-		// rows do not fill the viewport. That prevents a stare at blank
-		// space below a sparse month until the user navigates.
-		var cmds []tea.Cmd
-		if m.viewMode == viewAgenda {
-			if cmd := m.agenda.MaybeFillViewport(); cmd != nil {
-				cmds = append(cmds, cmd)
-			}
-		}
-		if m.pendingOpenEvent.ID != 0 {
-			matched := matchOpenEvent(m.events, m.pendingOpenEvent)
-			m.agenda = m.agenda.SelectEvent(matched.ID, matched.StartTime)
-			openEv := matched
-			m.pendingOpenEvent = event.Event{}
-			cmds = append(cmds, func() tea.Msg { return EventViewRequestedMsg{Event: openEv} })
-		}
-		return m, tea.Batch(cmds...)
-
-	case calendarsLoadedMsg:
-		if msg.err == nil {
-			m.calendars = msg.calendars
-			m.accounts = msg.accounts
-			// Overlay any unconfirmed reorder so a reload that races an
-			// in-flight SetOrder reflects the user's move instead of the stale
-			// DB order. Applied to m.calendars itself so the sidebar items and
-			// the manage dialog (both built below from m.calendars) agree.
-			for id, pos := range m.pendingOrder {
-				if c, ok := m.calendars[id]; ok {
-					c.DisplayOrder = pos
-					m.calendars[id] = c
-				} else {
-					// Gone from a full reload (deleted) — drop the stale pending
-					// entry so the map can't grow without bound across
-					// reorder+delete cycles.
-					delete(m.pendingOrder, id)
-				}
-			}
-			for accountID, pos := range m.pendingAccountOrder {
-				found := false
-				for id, c := range m.calendars {
-					if c.AccountID != accountID {
-						continue
-					}
-					found = true
-					c.AccountOrder = pos
-					m.calendars[id] = c
-				}
-				if !found {
-					delete(m.pendingAccountOrder, accountID)
-				}
-			}
-			m.sidebar = m.sidebar.SetList(m.sidebar.List().SetItems(sortedCalendarListItems(m.calendars)))
-			// Prune stale hidden IDs after CalendarListModel has done its pruning.
-			m.hiddenCalendars = m.sidebar.List().HiddenSet()
-			m.saveUIState()
-			// Rebuild the per-view CalendarEvent slices so rename/color edits
-			// reflect immediately — eventsToCalendar reads colors from
-			// m.calendars at conversion time.
-			m = m.refreshCalendarViews()
-			if m.calendarManagerOpen {
-				m.calendarManager = m.calendarManager.SetData(m.calendars, m.hiddenCalendars)
-				if m.calendarManager.Screen() == CalendarManagerScreenAccount {
-					if params, ok := m.accountSettingsParams(m.calendarManager.ActiveAccountID()); ok {
-						m.calendarManager = m.calendarManager.OpenAccount(params).SetSize(m.width, m.height)
-					}
-				}
-			}
-		}
-		return m, nil
-
-	case CalendarMonthChangedMsg:
-		return m, m.loadEvents()
-
-	case WeekChangedMsg:
-		return m, m.loadEvents()
-
-	case DayChangedMsg:
-		return m, m.loadEvents()
-
-	case AgendaCursorChangedMsg:
-		m.agenda = m.agenda.ResetWindow(msg.Day)
-		return m, m.loadEvents()
-
-	case AgendaReloadMsg:
-		return m, m.loadEventsIncremental()
-
-	case AgendaEmptyDaysToggledMsg:
-		m.saveUIState()
-		return m, nil
-
-	case CalendarDaySelectedMsg:
-		dayEvents := eventsOn(filterVisibleEvents(m.events, m.hiddenCalendars), msg.Day)
-		if m.clickedEventID > 0 {
-			clicked := m.clickedEventID
-			m.clickedEventID = 0
-			for _, e := range dayEvents {
-				if e.ID == clicked {
-					ev := e
-					return m, func() tea.Msg { return EventViewRequestedMsg{Event: ev} }
-				}
-			}
-		}
-		m.dialog = NewEventDialogModel(msg.Day, dayEvents, m.calendars, newThemedHelp(m.theme)).
-			SetSelectedColor(m.theme.Selected).
-			SetSize(m.width, m.height)
-		m.dialogOpen = true
-		return m, nil
-
-	case EventCreateMsg:
-		form, cmd := NewEventFormModel(msg.Day, eventFormCalendars(m.calendars), m.theme)
-		return m.openEventForm(form, cmd)
-
-	case EventEditMsg:
-		if cmd, blocked := m.blockReadOnlyCalendarMutation(msg.Event.CalendarID); blocked {
-			return m, cmd
-		}
-		ev := msg.Event
-		// ev.StartTime carries the clicked occurrence's time for recurring
-		// events (queryEventsRange overwrites it with InstanceTime). The
-		// fresh Get below returns the master row, which has the original
-		// DTSTART. Capture the clicked instance time first. Pass it
-		// alongside so the form can prompt for scope on save.
-		instanceTime := ev.StartTime
-		instanceEnd := ev.EndTime
-		return m, func() tea.Msg {
-			ctx := context.Background()
-			fresh, err := m.app.Events.Get(ctx, ev.ID)
-			if err != nil {
-				return eventEditLoadedMsg{err: err}
-			}
-			attendees, err := m.app.Events.ListAttendees(ctx, ev.ID)
-			if err != nil {
-				return eventEditLoadedMsg{err: err}
-			}
-			fresh.Attendees = attendees
-			// Load the alarms too. A save writes the form list over the
-			// stored rows, so an empty list deletes every alarm of the
-			// event, including a preserved sync-only one (issue #579).
-			alarms, err := m.app.Events.ListAlarms(ctx, ev.ID)
-			if err != nil {
-				return eventEditLoadedMsg{err: err}
-			}
-			fresh.Alarms = alarms
-			// For a recurring instance, overwrite the master's DTSTART with
-			// the clicked occurrence. The form's date/time fields then
-			// reflect what the user actually clicked, not the original
-			// series start.
-			if fresh.RecurrenceRule != "" && !instanceTime.IsZero() {
-				fresh.StartTime = instanceTime
-				if !instanceEnd.IsZero() {
-					fresh.EndTime = instanceEnd
-				}
-				return eventEditLoadedMsg{event: fresh, instanceTime: instanceTime}
-			}
-			return eventEditLoadedMsg{event: fresh}
-		}
-
-	case eventEditLoadedMsg:
-		if msg.err != nil {
-			m.err = msg.err
-			return m, nil
-		}
-		form, cmd := NewEventFormModelForEditInstance(msg.event, msg.instanceTime, eventFormCalendars(m.calendars), m.theme)
-		m.dialogOpen = false
-		if m.viewDialogOpen {
-			m.viewReturnEvent = msg.event
-		}
-		m.viewDialogOpen = false
-		return m.openEventForm(form, cmd)
-
-	case EventViewRequestedMsg:
-		ev := msg.Event
-		return m, func() tea.Msg {
-			ctx := context.Background()
-			fresh, err := m.app.Events.Get(ctx, ev.ID)
-			if err != nil {
-				return eventViewLoadedMsg{err: err}
-			}
-			attendees, err := m.app.Events.ListAttendees(ctx, ev.ID)
-			if err != nil {
-				return eventViewLoadedMsg{err: err}
-			}
-			fresh.Attendees = attendees
-			alarms, err := m.app.Events.ListAlarms(ctx, ev.ID)
-			if err != nil {
-				return eventViewLoadedMsg{err: err}
-			}
-			fresh.Alarms = alarms
-			fresh = applyOccurrenceTimes(fresh, ev)
-			return eventViewLoadedMsg{event: fresh}
-		}
-
-	case eventViewLoadedMsg:
-		if msg.err != nil {
-			m.err = msg.err
-			return m, nil
-		}
-		cal := m.calendars[msg.event.CalendarID]
-		m.viewDialog = NewEventViewDialogModel(msg.event, cal, m.theme).
-			SetSize(m.width, m.height)
-		m.viewDialogOpen = true
-		return m, nil
-
-	case EventViewClosedMsg:
-		m.viewDialogOpen = false
-		return m, nil
-
-	case EventDuplicateMsg:
-		form, cmd := NewEventFormModelForDuplicate(msg.Event, eventFormCalendars(m.calendars), m.theme)
-		if m.viewDialogOpen {
-			m.viewReturnEvent = msg.Event
-		}
-		m.viewDialogOpen = false
-		return m.openEventForm(form, cmd)
-
-	case EventFormSaveMsg:
-		if cmd, blocked := m.blockReadOnlyCalendarMutation(msg.CalendarID); blocked {
-			return m, cmd
-		}
-		// editID is read from the live form model, not the message. The
-		// form's OnSubmit closure is bound before NewEventFormModelForEdit
-		// assigns editID, so EventFormSaveMsg cannot carry that value
-		// reliably — see event_form.go:EventFormSaveMsg for the rationale.
-		editID := m.form.editID
-		m.formOpen = false
-		attendees := msg.Attendees
-		alarms := msg.Alarms
-		// Editing one occurrence of a recurring series → defer the actual
-		// write until the user picks a scope. The dialog dispatch
-		// (ChoiceDialogResultMsg below) reads m.pendingEditSave and routes
-		// to UpdateInstance / UpdateFromInstance / Update.
-		if editID > 0 && !msg.InstanceTime.IsZero() {
-			m.pendingEditSave = msg
-			m.pendingScopeKind = pendingScopeEdit
-			m.choiceDialog = NewChoiceDialogModel(
-				fmt.Sprintf("Update %q?", msg.Title),
-				m.theme,
-				"This event", "This and following", "All events",
-			).SetSize(m.width, m.height)
-			m.choiceOpen = true
-			return m, nil
-		}
-		if editID > 0 {
-			eventID := editID
-			calID := msg.CalendarID
-			return m, func() tea.Msg {
-				ctx := context.Background()
-				// Update the row and its attendees/alarms in one transaction so a
-				// failed child write rolls the whole edit back (issue #87).
-				_, err := m.app.Events.UpdateWithRelations(ctx, eventID, event.UpdateParams{
-					CalendarID:     calID,
-					Title:          msg.Title,
-					Description:    msg.Description,
-					Location:       msg.Location,
-					ConferenceURI:  msg.ConferenceURI,
-					StartTime:      msg.StartTime,
-					EndTime:        msg.EndTime,
-					AllDay:         msg.AllDay,
-					RecurrenceRule: msg.RecurrenceRule,
-					Timezone:       msg.Timezone,
-					Transp:         msg.Transp,
-					Class:          msg.Class,
-					Categories:     msg.Categories,
-				}, attendees, alarms)
-				return eventUpdatedMsg{calendarID: calID, err: err}
-			}
-		}
-		calID := msg.CalendarID
-		return m, func() tea.Msg {
-			ctx := context.Background()
-			created, err := m.app.Events.Create(ctx, event.CreateParams{
-				CalendarID:     calID,
-				Title:          msg.Title,
-				Description:    msg.Description,
-				Location:       msg.Location,
-				ConferenceURI:  msg.ConferenceURI,
-				StartTime:      msg.StartTime,
-				EndTime:        msg.EndTime,
-				AllDay:         msg.AllDay,
-				RecurrenceRule: msg.RecurrenceRule,
-				Timezone:       msg.Timezone,
-				Transp:         msg.Transp,
-				Class:          msg.Class,
-				Categories:     msg.Categories,
-			})
-			if err != nil {
-				return eventCreatedMsg{calendarID: calID, err: err}
-			}
-			if len(attendees) > 0 {
-				if err = m.app.Events.ReplaceAttendees(ctx, created.ID, attendees); err != nil {
-					return eventCreatedMsg{calendarID: calID, err: err}
-				}
-			}
-			if len(alarms) > 0 {
-				err = m.app.Events.ReplaceAlarms(ctx, created.ID, alarms)
-			}
-			return eventCreatedMsg{calendarID: calID, err: err}
-		}
-
-	case EventFormClosedMsg:
-		m.formOpen = false
-		if m.viewReturnEvent.ID != 0 {
-			ev := m.viewReturnEvent
-			m.viewReturnEvent = event.Event{}
-			return m, func() tea.Msg { return EventViewRequestedMsg{Event: ev} }
-		}
-		return m, nil
-
-	case PaletteSelectedMsg:
-		m.paletteOpen = false
-		if msg.Action == nil {
-			return m, nil
-		}
-		action := msg.Action
-		return m, func() tea.Msg { return action() }
-
-	case PaletteClosedMsg:
-		m.paletteOpen = false
-		return m, nil
-
-	case SwitchViewMsg:
-		return m.switchToView(msg.Mode)
-
-	case GoToTodayMsg:
-		return m.goToToday()
-
-	case eventUpdateAfterScopeMsg:
-		if msg.err != nil {
-			m.err = msg.err
-			m.viewReturnEvent = event.Event{}
-			return m, nil
-		}
-		cmds := []tea.Cmd{m.loadEvents()}
-		if push := m.runOpportunisticPush(msg.calendarID); push != nil {
-			cmds = append(cmds, push)
-		}
-		// On scope-dispatched edits we don't return to the prior view dialog;
-		// the underlying row may have been replaced (e.g. "This and following"
-		// creates a new master with a fresh UID).
-		m.viewReturnEvent = event.Event{}
-		return m, tea.Batch(cmds...)
-
-	case ToggleSidebarMsg:
-		return m.toggleSidebar()
-
-	case ToggleWeekNumbersMsg:
-		return m.toggleWeekNumbers()
-
-	case ToggleWeekStartMsg:
-		return m.toggleWeekStart()
-
-	case eventCreatedMsg:
-		if msg.err != nil {
-			m.err = msg.err
-			m.viewReturnEvent = event.Event{}
-			return m, nil
-		}
-		cmds := []tea.Cmd{m.loadEvents()}
-		if push := m.runOpportunisticPush(msg.calendarID); push != nil {
-			cmds = append(cmds, push)
-		}
-		if m.viewReturnEvent.ID != 0 {
-			ev := m.viewReturnEvent
-			m.viewReturnEvent = event.Event{}
-			cmds = append(cmds, func() tea.Msg { return EventViewRequestedMsg{Event: ev} })
-		}
-		return m, tea.Batch(cmds...)
-
-	case eventUpdatedMsg:
-		if msg.err != nil {
-			m.err = msg.err
-			m.viewReturnEvent = event.Event{}
-			return m, nil
-		}
-		cmds := []tea.Cmd{m.loadEvents()}
-		if push := m.runOpportunisticPush(msg.calendarID); push != nil {
-			cmds = append(cmds, push)
-		}
-		if m.viewReturnEvent.ID != 0 {
-			ev := m.viewReturnEvent
-			m.viewReturnEvent = event.Event{}
-			cmds = append(cmds, func() tea.Msg { return EventViewRequestedMsg{Event: ev} })
-		}
-		return m, tea.Batch(cmds...)
-
-	case EventRSVPMsg:
-		ev := msg.Event
-		ownerEmail := m.calendars[ev.CalendarID].OwnerEmail
-		return m, func() tea.Msg {
-			ctx := context.Background()
-			attendees, err := m.app.Events.ListAttendees(ctx, ev.ID)
-			if err != nil {
-				return eventRSVPUpdatedMsg{err: err}
-			}
-			for i, att := range attendees {
-				if strings.EqualFold(att.Email, ownerEmail) {
-					attendees[i].RSVPStatus = msg.Status
-					break
-				}
-			}
-			if err := m.app.Events.ReplaceAttendees(ctx, ev.ID, attendees); err != nil {
-				return eventRSVPUpdatedMsg{err: err}
-			}
-			ev.Attendees = attendees
-			return eventRSVPUpdatedMsg{event: ev}
-		}
-
-	case eventRSVPUpdatedMsg:
-		if msg.err != nil {
-			m.err = msg.err
-			return m, nil
-		}
-		// Rebuild the open view dialog so the user sees their new RSVP
-		// status immediately; loadEvents only repaints the grid behind it.
-		if m.viewDialogOpen && m.viewDialog.event.ID == msg.event.ID {
-			cal := m.calendars[msg.event.CalendarID]
-			m.viewDialog = NewEventViewDialogModel(msg.event, cal, m.theme).
-				SetSize(m.width, m.height)
-		}
-		return m, m.loadEvents()
-
-	case DialogDayChangedMsg:
-		if m.viewMode == viewDay {
-			prevDay := m.day.cursor.Format("2006-01-02")
-			m.day.cursor = msg.Day
-			if m.day.cursor.Format("2006-01-02") != prevDay {
-				m.dialog = NewEventDialogModel(msg.Day, nil, m.calendars, newThemedHelp(m.theme)).
-					SetSelectedColor(m.theme.Selected).
-					SetSize(m.width, m.height)
-				return m, m.loadEvents()
-			}
-			dayEvents := eventsOn(filterVisibleEvents(m.events, m.hiddenCalendars), msg.Day)
-			m.dialog = NewEventDialogModel(msg.Day, dayEvents, m.calendars, newThemedHelp(m.theme)).
-				SetSelectedColor(m.theme.Selected).
-				SetSize(m.width, m.height)
-			return m, nil
-		}
-		if m.viewMode == viewWeek {
-			prevWeek := m.week.WeekStartDate()
-			m.week.cursor = msg.Day
-			if m.week.WeekStartDate() != prevWeek {
-				m.dialog = NewEventDialogModel(msg.Day, nil, m.calendars, newThemedHelp(m.theme)).
-					SetSelectedColor(m.theme.Selected).
-					SetSize(m.width, m.height)
-				return m, m.loadEvents()
-			}
-			dayEvents := eventsOn(filterVisibleEvents(m.events, m.hiddenCalendars), msg.Day)
-			m.dialog = NewEventDialogModel(msg.Day, dayEvents, m.calendars, newThemedHelp(m.theme)).
-				SetSelectedColor(m.theme.Selected).
-				SetSize(m.width, m.height)
-			return m, nil
-		}
-		m.calendar.cursor = msg.Day
-		if msg.Day.Year() != m.calendar.month.Year() || msg.Day.Month() != m.calendar.month.Month() {
-			m.calendar.month = time.Date(msg.Day.Year(), msg.Day.Month(), 1, 0, 0, 0, 0, msg.Day.Location())
-			m.dialog = NewEventDialogModel(msg.Day, nil, m.calendars, newThemedHelp(m.theme)).
-				SetSelectedColor(m.theme.Selected).
-				SetSize(m.width, m.height)
-			return m, m.loadEvents()
-		}
-		dayEvents := eventsOn(filterVisibleEvents(m.events, m.hiddenCalendars), msg.Day)
-		m.dialog = NewEventDialogModel(msg.Day, dayEvents, m.calendars, newThemedHelp(m.theme)).
-			SetSelectedColor(m.theme.Selected).
-			SetSize(m.width, m.height)
-		return m, nil
-
-	case EventDialogClosedMsg:
-		m.dialogOpen = false
-		return m, nil
-
-	case EventDeleteMsg:
-		if cmd, blocked := m.blockReadOnlyCalendarMutation(msg.Event.CalendarID); blocked {
-			return m, cmd
-		}
-		m.pendingDelete = msg.Event
-		if msg.Event.RecurrenceRule != "" {
-			m.pendingScopeKind = pendingScopeDelete
-			m.choiceDialog = NewChoiceDialogModel(
-				fmt.Sprintf("Delete %q?", msg.Event.Title),
-				m.theme,
-				"This event", "This and following", "All events",
-			).SetSize(m.width, m.height)
-			m.choiceOpen = true
-		} else {
-			m.confirmDialog = NewConfirmDialogModel(
-				fmt.Sprintf("Delete %q?", msg.Event.Title),
-				"Delete",
-				m.theme,
-			).Destructive().SetSize(m.width, m.height)
-			m.confirmOpen = true
-		}
-		return m, nil
-
-	case SidebarFocusEscapedMsg:
-		m.sidebar = m.sidebar.Blur()
-		m.focus = focusCalendar
-		return m, nil
-
-	case MiniMonthDateSelectedMsg:
-		m = m.navigateMainTo(msg.Date)
-		return m, m.loadEvents()
-
-	case CalendarVisibilityToggledMsg:
-		// Mirror the toggle into the open calendar manager. Its root list
-		// and reopened detail params then never go stale. The manager's own
-		// updateCalendar mirror is only reachable through this forward.
-		if m.calendarManagerOpen {
-			m.calendarManager, _ = m.calendarManager.Update(msg)
-		}
-		if m.hiddenCalendars == nil {
-			m.hiddenCalendars = map[int64]bool{}
-		}
-		if msg.Hidden {
-			m.hiddenCalendars[msg.ID] = true
-		} else {
-			delete(m.hiddenCalendars, msg.ID)
-		}
-		list := m.sidebar.List().SetHidden(msg.ID, msg.Hidden)
-		m.sidebar = m.sidebar.SetList(list)
-		m.saveUIState()
-		m = m.refreshCalendarViews()
-		if m.dialogOpen {
-			dayEvents := eventsOn(filterVisibleEvents(m.events, m.hiddenCalendars), m.dialog.day)
-			m.dialog = m.dialog.SetEvents(dayEvents)
-		}
-		return m, nil
-
-	case CalendarReorderedMsg:
-		// A reorder can originate from the sidebar list or the manage
-		// dialog. Mirror it into m.calendars so both views (and any later
-		// reload-from-cache) stay coherent. Record it as pending. A
-		// reload that races the async SetOrder below then does not revert
-		// to the stale DB order.
-		ids := msg.IDs
-		if m.pendingOrder == nil {
-			m.pendingOrder = make(map[int64]int64, len(ids))
-		}
-		for i, id := range ids {
-			m.pendingOrder[id] = int64(i)
-			if info, ok := m.calendars[id]; ok {
-				info.DisplayOrder = int64(i)
-				m.calendars[id] = info
-			}
-		}
-		// Re-sort the sidebar from the updated order. A dialog-originated
-		// reorder then shows up behind the dialog. For a sidebar-originated
-		// one it re-applies the order the list already swapped to. Keep
-		// the cursor by calendar identity. A reorder from either surface
-		// then never leaves the sidebar highlight on a different calendar.
-		m.sidebar = m.sidebar.SetList(m.sidebar.List().SetItemsPreservingCursor(sortedCalendarListItems(m.calendars)))
-		if m.calendarManagerOpen {
-			m.calendarManager = m.calendarManager.SetData(m.calendars, m.hiddenCalendars)
-		}
-		return m, func() tea.Msg {
-			return calendarOrderSavedMsg{ids: ids, err: m.app.Calendars.SetOrder(context.Background(), ids)}
-		}
-
-	case AccountReorderedMsg:
-		ids := slices.Clone(msg.IDs)
-		m.pendingAccountOrder = make(map[int64]int64, len(ids))
-		m.pendingAccountOrderIDs = ids
-		for i, accountID := range ids {
-			position := int64(i)
-			m.pendingAccountOrder[accountID] = position
-			for calendarID, info := range m.calendars {
-				if info.AccountID == accountID {
-					info.AccountOrder = position
-					m.calendars[calendarID] = info
-				}
-			}
-		}
-		m.sidebar = m.sidebar.SetList(
-			m.sidebar.List().SetItemsPreservingCursor(sortedCalendarListItems(m.calendars)),
-		)
-		if m.accountOrderSaveInFlight {
-			return m, nil
-		}
-		return m, m.beginAccountOrderSave(ids)
-
-	case accountOrderSavedMsg:
-		m.accountOrderSaveInFlight = false
-		latest := m.pendingAccountOrderIDs
-		if len(latest) > 0 && !slices.Equal(msg.ids, latest) {
-			saveLatest := m.beginAccountOrderSave(latest)
-			if msg.err != nil {
-				return m, tea.Batch(m.toast.Failed(msg.err.Error()), saveLatest)
-			}
-			return m, saveLatest
-		}
-		for i, id := range msg.ids {
-			if m.pendingAccountOrder[id] == int64(i) {
-				delete(m.pendingAccountOrder, id)
-			}
-		}
-		m.pendingAccountOrderIDs = nil
-		if msg.err != nil {
-			return m, tea.Batch(m.toast.Failed(msg.err.Error()), m.loadCalendars())
-		}
-		return m, nil
-
-	case calendarOrderSavedMsg:
-		if msg.err != nil {
-			return m, m.toast.Failed(msg.err.Error())
-		}
-		// Clear only the positions this save confirmed. A newer reorder may
-		// have updated m.pendingOrder while this write was in flight. That
-		// one stays pending until its own save lands.
-		for i, id := range msg.ids {
-			if m.pendingOrder[id] == int64(i) {
-				delete(m.pendingOrder, id)
-			}
-		}
-		return m, nil
-	case AccountRenameRequestedMsg:
-		if m.syncing || !m.accountRenameOpen {
-			return m, nil
-		}
-		m.accountRenameOpen = false
-		m.syncing = true
-		m.syncStatus = "Renaming account…"
-		request := msg
-		return m, func() tea.Msg {
-			renamed, err := m.app.Accounts.Rename(context.Background(), request.AccountID, request.Name)
-			return accountRenameFinishedMsg{account: renamed, err: err}
-		}
-
-	case accountRenameCancelledMsg:
-		m.accountRenameOpen = false
-		m.accountRenameFromSettings = false
-		return m, nil
-
-	case AccountSettingsRequestedMsg:
-		params, ok := m.accountSettingsParams(msg.AccountID)
-		if !ok {
-			return m, nil
-		}
-		m.calendarManager = m.calendarManager.OpenAccount(params).SetSize(m.width, m.height)
-		m.calendarManagerOpen = true
-		return m, nil
-
-	case AccountSettingsManageRequestedMsg:
-		if m.syncing || m.oauthFlowOpen || m.oauthPending ||
-			!m.calendarManagerOpen || m.calendarManager.ActiveAccountID() != msg.AccountID {
-			return m, nil
-		}
-		if _, ok := m.accounts[msg.AccountID]; !ok {
-			return m, nil
-		}
-		m.calendarManagerOpen = true
-		m.accountManagementGeneration++
-		m.pendingAccountManagementID = msg.AccountID
-		m.syncing = true
-		m.syncStatus = "Discovering calendars…"
-		return m, tea.Batch(
-			m.syncSpinner.Tick,
-			m.discoverAccountCalendars(msg.AccountID, m.accountManagementGeneration),
-		)
-
-	case AccountSettingsSyncRequestedMsg:
-		if m.syncing || !m.calendarManagerOpen ||
-			m.calendarManager.ActiveAccountID() != msg.AccountID {
-			return m, nil
-		}
-		params, ok := m.accountSettingsParams(msg.AccountID)
-		if !ok {
-			return m, nil
-		}
-		m.syncing = true
-		m.syncStatus = "Syncing " + textsafe.Display(params.DisplayName) + "…"
-		return m, tea.Batch(m.syncSpinner.Tick, m.runSyncAccount(msg.AccountID, params.DisplayName))
-
-	case AccountSettingsRenameRequestedMsg:
-		if m.syncing || !m.calendarManagerOpen ||
-			m.calendarManager.ActiveAccountID() != msg.AccountID {
-			return m, nil
-		}
-		configured, ok := m.accounts[msg.AccountID]
-		if !ok {
-			return m, nil
-		}
-		m.accountRename = newAccountRenameDialogModel(configured, m.theme).
-			SetSize(m.width, m.height)
-		m.accountRenameFromSettings = true
-		m.accountRenameOpen = true
-		return m, nil
-
-	case AccountOAuthConfigSubmittedMsg:
-		if m.oauthFlowOpen || m.oauthPending || !m.accountOAuthConfigOpen ||
-			m.accountOAuthConfig.accountID != msg.AccountID {
-			return m, nil
-		}
-		configured, ok := m.accounts[msg.AccountID]
-		if !ok || !calendarAuthIsOAuth(configured.AuthType) {
-			return m, nil
-		}
-		m.accountOAuthConfigOpen = false
-		m.oauthPending = true
-		m.syncStatus = "Preparing sign-in…"
-		return m, m.prepareAccountReauth(configured, msg.ClientID, msg.ClientSecret)
-
-	case AccountOAuthConfigClosedMsg:
-		if !m.accountOAuthConfigOpen || m.accountOAuthConfig.accountID != msg.AccountID {
-			return m, nil
-		}
-		m.accountOAuthConfigOpen = false
-		m.oauthPending = false
-		if params, ok := m.accountSettingsParams(msg.AccountID); ok {
-			m.calendarManager = m.calendarManager.OpenAccount(params).SetSize(m.width, m.height)
-			m.calendarManagerOpen = true
-		}
-		m.statusToken++
-		m.syncStatus = "Authorization cancelled"
-		return m, m.expireStatusAfter(6*time.Second, m.statusToken)
-
-	case AccountSettingsReauthRequestedMsg:
-		if m.syncing || m.oauthFlowOpen || m.oauthPending ||
-			!m.calendarManagerOpen || m.calendarManager.ActiveAccountID() != msg.AccountID {
-			return m, nil
-		}
-		configured, ok := m.accounts[msg.AccountID]
-		if !ok || !calendarAuthIsOAuth(configured.AuthType) {
-			return m, nil
-		}
-		m.calendarManagerOpen = false
-		m.oauthPending = true
-		m.syncStatus = "Preparing sign-in…"
-		return m, m.prepareAccountReauth(configured, "", "")
-
-	case AccountSettingsUpdateCredentialsRequestedMsg:
-		if m.syncing || m.oauthFlowOpen || m.oauthPending ||
-			!m.calendarManagerOpen || m.calendarManager.ActiveAccountID() != msg.AccountID {
-			return m, nil
-		}
-		configured, ok := m.accounts[msg.AccountID]
-		if !ok || !accountAuthIsBasicOrBearer(configured.AuthType) {
-			return m, nil
-		}
-		m.calendarManagerOpen = false
-		m.accountCredentials = NewAccountCredentialsDialogModel(
-			configured.ID, configured.DisplayName, configured.AuthType, configured.Username, m.theme,
-		).SetSize(m.width, m.height)
-		m.accountCredentialsOpen = true
-		return m, nil
-
-	case AccountCredentialsUpdateSubmittedMsg:
-		if m.syncing || m.oauthFlowOpen || m.oauthPending || !m.accountCredentialsOpen ||
-			m.accountCredentials.accountID != msg.AccountID {
-			return m, nil
-		}
-		configured, ok := m.accounts[msg.AccountID]
-		if !ok || !accountAuthIsBasicOrBearer(configured.AuthType) {
-			return m, nil
-		}
-		m.accountCredentialsOpen = false
-		m.syncing = true
-		m.syncStatus = "Updating credentials…"
-		return m, tea.Batch(
-			m.syncSpinner.Tick,
-			m.updateAccountCredentials(configured, msg.Secret),
-		)
-
-	case AccountCredentialsUpdateClosedMsg:
-		if !m.accountCredentialsOpen || m.accountCredentials.accountID != msg.AccountID {
-			return m, nil
-		}
-		m.accountCredentialsOpen = false
-		m.statusToken++
-		m.syncStatus = "Credential update cancelled"
-		m = m.reopenAccountSettings(msg.AccountID)
-		return m, m.expireStatusAfter(6*time.Second, m.statusToken)
-
-	case AccountSettingsRemoveRequestedMsg:
-		if m.syncing || !m.calendarManagerOpen ||
-			m.calendarManager.ActiveAccountID() != msg.AccountID {
-			return m, nil
-		}
-		params, ok := m.accountSettingsParams(msg.AccountID)
-		if !ok {
-			return m, nil
-		}
-		name := textsafe.Display(params.DisplayName)
-		message := fmt.Sprintf(
-			"Remove account “%s” from Chroncal?\n\n%d downloaded %s will be kept as local calendars.\nRemote links and stored sign-in will be removed.",
-			name,
-			params.CalendarCount,
-			pluralize(params.CalendarCount, "calendar", "calendars"),
-		)
-		if draft := m.calendarManager.LocalDraft(); m.calendarManagerOpen &&
-			draft != nil && draft.AccountID == msg.AccountID {
-			message += "\nAny unsaved calendar edits will be discarded."
-		}
-		m.pendingAccountRemoveID = msg.AccountID
-		m.pendingAccountRemoveName = name
-		m.confirmDialog = NewConfirmDialogModel(message, "Remove Account", m.theme).
-			Destructive().
-			SetSize(m.width, m.height)
-		m.confirmOpen = true
-		return m, nil
-
-	case AccountSettingsClosedMsg:
-		if m.syncing || m.oauthPending {
-			return m, nil
-		}
-		if m.calendarManagerOpen {
-			m.calendarManager = m.calendarManager.CloseAccount()
-		}
-		return m, nil
-
-	case CalendarDiscoveryRequestedMsg:
-		if m.syncing || m.oauthFlowOpen {
-			return m, nil
-		}
-		if calendarAuthIsOAuth(msg.AuthType) {
-			m.oauthPurpose = oauthFlowPurpose{
-				calendarDiscovery:    true,
-				calendarDiscoveryMsg: msg,
-			}
-			return m.startOAuthFlow(msg.OAuthClientID, msg.OAuthClientSecret)
-		}
-		cred := auth.Credential{Username: msg.Username}
-		if msg.AuthType == "bearer" {
-			cred.AccessToken = msg.Secret
-		} else {
-			cred.Password = msg.Secret
-		}
-		m.syncing = true
-		m.syncStatus = "Adding account…"
-		return m, tea.Batch(m.syncSpinner.Tick, m.connectAndDiscoverCalendar(msg, cred))
-
-	case accountDiscoveryReadyMsg:
-		m.syncing = false
-		if msg.err != nil {
-			m.calendarManagerOpen = true
-			m.statusToken++
-			m.syncStatus = "Couldn’t add account: " + msg.err.Error()
-			m.calendarManager = m.calendarManager.WithTestStatus(lipgloss.NewStyle().Foreground(m.theme.Error), "✗ "+msg.err.Error())
-			return m, m.expireStatusAfter(10*time.Second, m.statusToken)
-		}
-		m.pendingDiscoveryAccountID = msg.discovery.Account.ID
-		m.pendingDiscoveryCreated = msg.createdAccount
-		m.calendarManager = m.calendarManager.ShowDiscovery(msg.discovery).SetSize(m.width, m.height)
-		if msg.createdAccount {
-			paths := make([]string, 0, len(msg.discovery.Calendars))
-			for _, remote := range msg.discovery.Calendars {
-				if remote.Importable && !remote.Missing {
-					paths = append(paths, remote.Path)
-				}
-			}
-			m.calendarManagerOpen = false
-			m.syncing = true
-			m.syncStatus = "Importing and syncing all calendars…"
-			return m, tea.Batch(
-				m.syncSpinner.Tick,
-				m.importAndSyncAccountCalendars(paths),
-			)
-		}
-		m.calendarManagerOpen = true
-		m.syncStatus = ""
-		return m, m.loadCalendars()
-
-	case AccountCalendarPickerClosedMsg:
-		m.statusToken++
-		if m.calendarManager.ManagingAccountCalendars() {
-			m.calendarManager = m.calendarManager.HideDiscovery()
-			m.calendarManagerOpen = true
-			m.syncStatus = "Calendar discovery cancelled"
-			return m, m.expireStatusAfter(6*time.Second, m.statusToken)
-		}
-		m.calendarManagerOpen = false
-		if m.pendingDiscoveryCreated && m.pendingDiscoveryAccountID != 0 {
-			accountID := m.pendingDiscoveryAccountID
-			m.pendingDiscoveryAccountID = 0
-			m.pendingDiscoveryCreated = false
-			m.syncing = true
-			m.syncStatus = "Cancelling calendar discovery…"
-			return m, tea.Batch(m.syncSpinner.Tick, m.discardDiscoveryAccount(accountID))
-		}
-		m.pendingDiscoveryAccountID = 0
-		m.pendingDiscoveryCreated = false
-		m.syncStatus = "Calendar discovery cancelled"
-		return m, m.expireStatusAfter(6*time.Second, m.statusToken)
-
-	case AccountCalendarsImportRequestedMsg:
-		if m.calendarManager.DiscoveryPicker() == nil ||
-			msg.AccountID != m.calendarManager.DiscoveryPicker().discovery.Account.ID {
-			return m, m.toast.Failed("calendar selection no longer matches the open account")
-		}
-		m.calendarManagerOpen = false
-		m.syncing = true
-		m.syncStatus = "Importing and syncing selected calendars…"
-		return m, tea.Batch(m.syncSpinner.Tick, m.importAndSyncAccountCalendars(msg.Paths))
-
-	case AccountCalendarsReconcileRequestedMsg:
-		selection, candidates, err := m.prepareAccountCalendarSelection(msg)
-		if err != nil {
-			return m, m.toast.Failed(err.Error())
-		}
-		if len(selection.removed) == 0 {
-			m.calendarManagerOpen = true
-			m.syncing = true
-			m.syncStatus = "Adding and syncing selected calendars…"
-			return m, tea.Batch(
-				m.syncSpinner.Tick,
-				m.reconcileAndSyncAccountCalendars(selection),
-			)
-		}
-		if len(candidates) > 0 {
-			m.pendingAccountSelection = selection
-			m.pendingAccountDefaultCandidates = candidates
-			m.pendingScopeKind = pendingScopeAccountSelectionPromote
-			labels := make([]string, len(candidates))
-			for i, candidate := range candidates {
-				labels[i] = textsafe.Display(candidate.name)
-			}
-			m.choiceDialog = NewChoiceDialogModel(
-				"The default calendar is being removed.\n\nChoose a new default before saving these changes:",
-				m.theme,
-				labels...,
-			).SetSize(m.width, m.height)
-			m.choiceOpen = true
-			return m, nil
-		}
-		m = m.showAccountCalendarRemovalConfirmation(selection)
-		return m, nil
-
-	case calendarDiscoveryDiscardedMsg:
-		m.syncing = false
-		m.calendarManager = m.calendarManager.HideDiscovery()
-		m.statusToken++
-		if msg.err != nil {
-			m.syncStatus = "Calendar discovery cancelled; cleanup failed: " + msg.err.Error()
-		} else {
-			m.syncStatus = "Calendar discovery cancelled"
-		}
-		return m, tea.Batch(
-			m.loadCalendars(),
-			m.expireStatusAfter(8*time.Second, m.statusToken),
-		)
-
-	case accountImportFinishedMsg:
-		m.syncing = false
-		m.statusToken++
-		switch {
-		case msg.err != nil:
-			m.calendarManagerOpen = m.calendarManager.DiscoveryPicker() != nil
-			m.syncStatus = "Calendar import failed: " + msg.err.Error()
-		case msg.created == 0 && msg.existing == 0:
-			m.syncStatus = "No calendars selected"
-			if m.pendingDiscoveryCreated && m.pendingDiscoveryAccountID != 0 {
-				accountID := m.pendingDiscoveryAccountID
-				m.pendingDiscoveryAccountID = 0
-				m.pendingDiscoveryCreated = false
-				m.syncing = true
-				m.syncStatus = "Cancelling calendar discovery…"
-				return m, tea.Batch(m.syncSpinner.Tick, m.discardDiscoveryAccount(accountID))
-			}
-			m.pendingDiscoveryAccountID = 0
-			m.pendingDiscoveryCreated = false
-		case msg.syncErr != nil:
-			m.pendingDiscoveryAccountID = 0
-			m.pendingDiscoveryCreated = false
-			m.syncStatus = fmt.Sprintf("Imported %d calendar(s); first sync failed: %v", msg.created, msg.syncErr)
-		case msg.created == 0:
-			m.pendingDiscoveryAccountID = 0
-			m.pendingDiscoveryCreated = false
-			m.syncStatus = fmt.Sprintf("%d selected calendar(s) already imported", msg.existing)
-		default:
-			m.pendingDiscoveryAccountID = 0
-			m.pendingDiscoveryCreated = false
-			m.syncStatus = fmt.Sprintf("Imported and synced %d calendar(s)", msg.synced)
-		}
-		if msg.err == nil {
-			m.calendarManager = m.calendarManager.HideDiscovery()
-			m.syncStatus = appendImportWarnings(m.syncStatus, msg.warnings)
-		}
-		return m, tea.Batch(
-			m.loadCalendars(),
-			m.loadEvents(),
-			m.expireStatusAfter(10*time.Second, m.statusToken),
-		)
-
-	case accountSelectionFinishedMsg:
-		m.syncing = false
-		m.statusToken++
-		if msg.err != nil {
-			if msg.accountManagement {
-				m.calendarManagerOpen = m.calendarManager.DiscoveryPicker() != nil
-			} else {
-				m.calendarManagerOpen = m.calendarManager.DiscoveryPicker() != nil
-			}
-			m.syncStatus = "Calendar changes failed: " + msg.err.Error()
-			return m, m.expireStatusAfter(10*time.Second, m.statusToken)
-		}
-
-		m.pendingDiscoveryAccountID = 0
-		m.pendingDiscoveryCreated = false
-		if msg.accountManagement {
-			m.calendarManager = m.calendarManager.HideDiscovery()
-			m.calendarManagerOpen = !msg.accountRemoved
-		} else {
-			m.calendarManager = m.calendarManager.HideDiscovery()
-			m.calendarManagerOpen = !msg.removedCurrent && !msg.accountRemoved
-		}
-		if draft := m.calendarManager.LocalDraft(); draft != nil &&
-			slices.Contains(msg.removedIDs, draft.ID) {
-			m.calendarManagerOpen = false
-		}
-		switch {
-		case msg.accountRemoved:
-			m.syncStatus = fmt.Sprintf("Removed account and %d calendar(s) from Chroncal", msg.removed)
-		case msg.syncErr != nil:
-			m.syncStatus = fmt.Sprintf(
-				"Updated calendars: added %d, removed %d; first sync failed: %v",
-				msg.created, msg.removed, msg.syncErr,
-			)
-		default:
-			m.syncStatus = fmt.Sprintf(
-				"Updated calendars: added %d, removed %d",
-				msg.created, msg.removed,
-			)
-		}
-		m.syncStatus = appendImportWarnings(m.syncStatus, msg.warnings)
-		return m, tea.Batch(
-			m.loadCalendars(),
-			m.loadEvents(),
-			m.expireStatusAfter(10*time.Second, m.statusToken),
-		)
-
-	case CalendarExportRequestedMsg:
-		m.calendarTransferGeneration++
-		m.calendarManager = m.calendarManager.OpenExport(msg.ID, msg.Name, m.calendarTransferGeneration)
-		m.calendarManagerOpen = true
-		return m, nil
-
-	case CalendarTransferClosedMsg:
-		m.calendarTransferGeneration++
-		m.syncing = false
-		m.calendarManager = m.calendarManager.CloseTransfer()
-		return m, nil
-
-	case CalendarImportPreviewRequestedMsg:
-		transfer, ok := m.calendarManager.Transfer()
-		if !ok || transfer.Generation() != msg.Generation {
-			return m, nil
-		}
-		m.syncing = true
-		m.syncStatus = "Reading iCal file…"
-		path := msg.Path
-		return m, func() tea.Msg {
-			preview, err := icaltransfer.ParseFile(path)
-			return calendarImportPreviewReadyMsg{Generation: msg.Generation, Path: path, Preview: preview, Err: err}
-		}
-
-	case calendarImportPreviewReadyMsg:
-		transfer, ok := m.calendarManager.Transfer()
-		if !ok || transfer.Generation() != msg.Generation {
-			return m, nil
-		}
-		m.syncing = false
-		if msg.Err != nil {
-			next := transfer.WithError(msg.Err)
-			m.calendarManager = m.calendarManager.SetTransfer(next)
-			m.syncStatus = "iCal preview failed: " + msg.Err.Error()
-			return m, nil
-		}
-		if msg.Preview.Events+msg.Preview.Todos+msg.Preview.Journals == 0 {
-			err := fmt.Errorf("no importable VEVENT, VTODO, or VJOURNAL components")
-			if msg.Preview.FreeBusy > 0 {
-				err = fmt.Errorf("%w; VFREEBUSY is preview-only", err)
-			}
-			m.calendarManager = m.calendarManager.SetTransfer(transfer.WithError(err))
-			m.syncStatus = "iCal preview failed: " + err.Error()
-			return m, nil
-		}
-		next := transfer.WithPreview(msg.Path, msg.Preview, calendarImportDestinations(m.calendars, msg.Preview)).
-			SetSize(m.width, m.height)
-		m.calendarManager = m.calendarManager.SetTransfer(next)
-		m.syncStatus = ""
-		return m, nil
-
-	case CalendarImportRequestedMsg:
-		transfer, ok := m.calendarManager.Transfer()
-		if !ok || transfer.Generation() != msg.Generation {
-			return m, nil
-		}
-		m.syncing = true
-		m.syncStatus = "Importing iCal file…"
-		request := msg
-		return m, func() tea.Msg {
-			ctx := context.Background()
-			calendarID := request.CalendarID
-			created := false
-			if calendarID == 0 {
-				cal, err := m.app.Calendars.Create(ctx, request.NewName, request.NewColor, "")
-				if err != nil {
-					return calendarImportFinishedMsg{Generation: request.Generation, Err: fmt.Errorf("create destination calendar: %w", err)}
-				}
-				calendarID = cal.ID
-				created = true
-			}
-			cleanup := func(cause error) calendarImportFinishedMsg {
-				if !created {
-					return calendarImportFinishedMsg{Generation: request.Generation, CalendarID: calendarID, Err: cause}
-				}
-				if err := m.app.Calendars.Delete(ctx, calendarID); err != nil {
-					cause = errors.Join(cause, fmt.Errorf("remove incomplete calendar: %w", err))
-				}
-				return calendarImportFinishedMsg{Generation: request.Generation, Err: cause}
-			}
-			if err := icaltransfer.ValidateDestination(ctx, m.app, calendarID, request.Preview); err != nil {
-				return cleanup(err)
-			}
-			summary := icaltransfer.Import(ctx, m.app, calendarID, &request.Preview.Result)
-			if summary.Failed > 0 {
-				return cleanup(fmt.Errorf("%d component(s) failed to import", summary.Failed))
-			}
-			return calendarImportFinishedMsg{Generation: request.Generation, CalendarID: calendarID, Summary: summary}
-		}
-
-	case calendarImportFinishedMsg:
-		transfer, ok := m.calendarManager.Transfer()
-		if !ok || transfer.Generation() != msg.Generation {
-			return m, nil
-		}
-		m.syncing = false
-		if msg.Err != nil {
-			if transfer, ok := m.calendarManager.Transfer(); ok {
-				m.calendarManager = m.calendarManager.SetTransfer(transfer.WithError(msg.Err))
-			}
-			m.syncStatus = "iCal import failed: " + msg.Err.Error()
-			return m, nil
-		}
-		m.calendarManager = m.calendarManager.CompleteTransfer(msg.CalendarID)
-		imported := len(msg.Summary.Events) + len(msg.Summary.Todos) + len(msg.Summary.Journals)
-		m.syncStatus = fmt.Sprintf("Imported %d item(s) with %d warning(s)", imported, len(msg.Summary.Warnings))
-		m.statusToken++
-		return m, tea.Batch(m.loadCalendars(), m.loadEvents(), m.expireStatusAfter(10*time.Second, m.statusToken))
-
-	case CalendarExportWriteRequestedMsg:
-		transfer, ok := m.calendarManager.Transfer()
-		if !ok || transfer.Generation() != msg.Generation {
-			return m, nil
-		}
-		m.syncing = true
-		m.syncStatus = "Exporting calendar…"
-		request := msg
-		return m, func() tea.Msg {
-			summary, err := icaltransfer.ExportCalendarFile(
-				context.Background(), m.app, request.CalendarID, request.Name, request.Path,
-			)
-			return calendarExportFinishedMsg{Generation: request.Generation, Path: request.Path, Summary: summary, Err: err}
-		}
-
-	case calendarExportFinishedMsg:
-		transfer, ok := m.calendarManager.Transfer()
-		if !ok || transfer.Generation() != msg.Generation {
-			return m, nil
-		}
-		m.syncing = false
-		if msg.Err != nil {
-			if transfer, ok := m.calendarManager.Transfer(); ok {
-				m.calendarManager = m.calendarManager.SetTransfer(transfer.WithError(msg.Err))
-			}
-			m.syncStatus = "Calendar export failed: " + msg.Err.Error()
-			return m, nil
-		}
-		m.calendarManager = m.calendarManager.CloseTransfer()
-		m.syncStatus = fmt.Sprintf("Exported %d events, %d todos, %d journals to %s",
-			msg.Summary.Events, msg.Summary.Todos, msg.Summary.Journals, msg.Path)
-		m.statusToken++
-		return m, m.expireStatusAfter(10*time.Second, m.statusToken)
-
-	case CalendarManagerRequestedMsg:
-		m.calendarManager = m.calendarManager.SetData(m.calendars, m.hiddenCalendars)
-		m.calendarManagerOpen = true
-		switch msg.Target {
-		case CalendarManagerTargetRoot:
-			m.calendarManager = m.calendarManager.CloseDetail()
-		case CalendarManagerTargetCalendar:
-			m.calendarManager = m.calendarManager.CloseDetail()
-			m.calendarManager = m.calendarManager.SelectCalendar(msg.CalendarID)
-			if info, ok := m.calendars[msg.CalendarID]; ok {
-				params := CalendarDialogParams{
-					ID:            msg.CalendarID,
-					AccountID:     info.AccountID,
-					AccountName:   info.AccountName,
-					Name:          info.Name,
-					Color:         info.Color,
-					Description:   info.Description,
-					OwnerEmail:    info.OwnerEmail,
-					RemoteLinked:  info.AccountID != 0,
-					LastSyncAt:    info.LastSyncAt,
-					LastSyncError: info.LastSyncError,
-					IsDefault:     info.IsDefault,
-					Hidden:        m.hiddenCalendars[msg.CalendarID],
-				}
-				if params.AccountName == "" {
-					if configured, exists := m.accounts[info.AccountID]; exists {
-						params.AccountName = configured.DisplayName
-					}
-				}
-				m.calendarManager = m.calendarManager.OpenCalendar(params).SetSize(m.width, m.height)
-			}
-		case CalendarManagerTargetAccount:
-			m.calendarManager = m.calendarManager.CloseDetail()
-			if params, ok := m.accountSettingsParams(msg.AccountID); ok {
-				m.calendarManager = m.calendarManager.OpenAccount(params).SetSize(m.width, m.height)
-			}
-		case CalendarManagerTargetLocalCreate:
-			m.calendarManager = m.calendarManager.CloseDetail()
-			params := CalendarDialogParams{Color: "#a6e3a1", OfferDefault: len(m.calendars) > 0}
-			m.calendarManager = m.calendarManager.OpenCalendar(params).SetSize(m.width, m.height)
-		case CalendarManagerTargetAccountConnect:
-			m.calendarManager = m.calendarManager.CloseDetail()
-			m.calendarManager = m.calendarManager.OpenAccountConnection().SetSize(m.width, m.height)
-		case CalendarManagerTargetImport:
-			m.calendarManager = m.calendarManager.CloseDetail()
-			m.calendarTransferGeneration++
-			m.calendarManager = m.calendarManager.OpenImport(m.calendarTransferGeneration)
-		}
-		return m, nil
-
-	case CalendarSavedMsg:
-		// Metadata saves stay blocked while a sync or OAuth completion is
-		// writing to the calendars table. Never do this in silence. Surface
-		// the block as a form error. The editor and draft then stay intact
-		// and the user knows to retry.
-		if m.syncing || m.oauthPending {
-			return m, func() tea.Msg {
-				return calendarMutationDoneMsg{err: errors.New("sync in progress — try again in a moment")}
-			}
-		}
-		// Keep the dialog open until the mutation succeeds so we can
-		// show validation errors (e.g. duplicate name) on the form.
-		saved := msg
-		return m, func() tea.Msg {
-			ctx := context.Background()
-			var (
-				cal calendar.Calendar
-				err error
-			)
-			if saved.ID == 0 {
-				cal, err = m.app.Calendars.Create(ctx, saved.Name, saved.Color, saved.Description)
-			} else {
-				cal, err = m.app.Calendars.Update(ctx, saved.ID, saved.Name, saved.Color, saved.Description)
-			}
-			if err != nil {
-				return calendarMutationDoneMsg{err: err}
-			}
-
-			if err := m.app.Calendars.SetOwnerEmail(ctx, cal.ID, saved.OwnerEmail); err != nil {
-				return calendarMutationDoneMsg{err: err}
-			}
-
-			// MakeDefault only matters on create; edit-mode default moves
-			// through the dedicated CalendarSetDefaultRequestedMsg path so
-			// the rule stays in one place.
-			if saved.ID == 0 && saved.MakeDefault {
-				if derr := m.app.Calendars.SetDefault(ctx, cal.ID); derr != nil {
-					return calendarMutationDoneMsg{err: derr}
-				}
-			}
-
-			return calendarMutationDoneMsg{err: nil}
-		}
-
-	case oauthFlowStartedMsg:
-		var cmd tea.Cmd
-		m.oauthFlow, cmd = m.oauthFlow.Update(msg)
-		return m, cmd
-
-	case oauthFlowDoneMsg:
-		var cmd tea.Cmd
-		m.oauthFlow, cmd = m.oauthFlow.Update(msg)
-		switch m.oauthFlow.State() {
-		case OAuthFlowDone:
-			m.oauthFlowOpen = false
-			if m.oauthPurpose.calendarDiscovery {
-				m.syncing = true
-				m.syncStatus = "Authorized; discovering calendars…"
-				return m, tea.Batch(m.syncSpinner.Tick, m.finishOAuthCalendarDiscovery(msg.result))
-			}
-			m.oauthPending = true
-			return m, m.finishOAuthReauth(msg.result)
-		case OAuthFlowCancelled:
-			m.oauthFlowOpen = false
-			m.statusToken++
-			if m.oauthPurpose.calendarDiscovery {
-				m.calendarManagerOpen = true
-				m.syncStatus = "Authorization cancelled"
-				return m, m.expireStatusAfter(6*time.Second, m.statusToken)
-			}
-			m = m.reopenAccountSettings(m.oauthPurpose.accountID)
-			m.syncStatus = "Authorization cancelled"
-			return m, m.expireStatusAfter(6*time.Second, m.statusToken)
-		default:
-			// Failed: the modal stays up showing the error; esc closes it.
-			// Abort so the chained Waiting-phase context cancel func is
-			// invoked now rather than leaking until process exit (Wait's
-			// deferred Close already released the listener).
-			m.oauthFlow.Abort()
-		}
-		return m, cmd
-
-	case oauthFlowClosedMsg:
-		m.oauthFlowOpen = false
-		if m.oauthPurpose.calendarDiscovery {
-			m.calendarManagerOpen = true
-		} else {
-			m = m.reopenAccountSettings(m.oauthPurpose.accountID)
-		}
-		return m, nil
-
-	case CalendarMoveToAccountRequestedMsg:
-		return m.beginCalendarMove(msg)
-
-	case CalendarKeepLocalRequestedMsg:
-		// Keep the edit dialog open behind the confirm, mirroring the delete
-		// flow: cancelling returns to the editor with the draft intact.
-		m.pendingCalendarKeepLocal = msg.ID
-		message := fmt.Sprintf(
-			"Keep %q as a local calendar?\n\nSyncing with its account stops; every downloaded event stays on this device.\nRe-adding it later from Manage Calendars creates a separate copy.",
-			msg.Name,
-		)
-		m.confirmDialog = NewConfirmDialogModel(message, "Keep as Local", m.theme).
-			SetSize(m.width, m.height)
-		m.confirmOpen = true
-		return m, nil
-
-	case CalendarSetDefaultRequestedMsg:
-		id := msg.ID
-		return m, func() tea.Msg {
-			// keepEditor: the Set as Default button lives on the edit form;
-			// success must not pop the form and discard an unsaved draft.
-			return calendarMutationDoneMsg{err: m.app.Calendars.SetDefault(context.Background(), id), keepEditor: true}
-		}
-
-	case CalendarTestRequestedMsg:
-		req := msg
-		return m, func() tea.Msg {
-			ctx, cancel := context.WithTimeout(context.Background(), 15*time.Second)
-			defer cancel()
-
-			meta, err := caldav.VerifyCalendarURL(ctx, req.URL, req.Username, req.Password, req.AuthType, req.AllowInsecure)
-			if err != nil {
-				return CalendarTestResultMsg{Message: err.Error()}
-			}
-			message := "Connected"
-			if meta.DisplayName != "" {
-				message = fmt.Sprintf("Connected · %s", meta.DisplayName)
-			}
-			return CalendarTestResultMsg{OK: true, Message: message}
-		}
-
-	case CalendarDeleteRequestedMsg:
-		// Fetch the event count before the confirm dialog. The user then
-		// knows how many events will be deleted alongside the calendar.
-		// When the target is the current default, slot in a picker dialog
-		// before the destructive confirm. The user then chooses the
-		// replacement default. Never promote in silence.
-		id, name := msg.ID, msg.Name
-		info, known := m.calendars[id]
-		if known && info.IsDefault {
-			candidates := defaultPromotionCandidates(m.calendars, id)
-			if len(candidates) == 0 {
-				// Last calendar: service will return ErrLastCalendar; let
-				// the normal confirm flow surface the error verbatim.
-				return m, func() tea.Msg {
-					count, _ := m.app.Events.CountByCalendar(context.Background(), id)
-					return calendarDeleteCountMsg{id: id, name: name, eventCount: count}
-				}
-			}
-			m.pendingCalendarDelete = id
-			m.pendingCalendarDeleteName = name
-			m.pendingCalendarPromoteCands = make([]int64, len(candidates))
-			labels := make([]string, len(candidates))
-			for i, c := range candidates {
-				m.pendingCalendarPromoteCands[i] = c.id
-				labels[i] = c.name
-			}
-			m.pendingScopeKind = pendingScopeCalendarPromote
-			message := fmt.Sprintf("%q is the default calendar.\n\nChoose a new default before deleting it:", name)
-			m.choiceDialog = NewChoiceDialogModel(message, m.theme, labels...).
-				SetSize(m.width, m.height)
-			m.choiceOpen = true
-			return m, nil
-		}
-		return m, func() tea.Msg {
-			count, _ := m.app.Events.CountByCalendar(context.Background(), id)
-			return calendarDeleteCountMsg{id: id, name: name, eventCount: count}
-		}
-
-	case calendarDeleteCountMsg:
-		// Keep the edit dialog open behind the confirm. If the user
-		// cancels the confirm, they return to the edit dialog. They do
-		// not lose their in-progress changes. The confirm dialog takes
-		// input priority, so the edit dialog is visible but inert.
-		m.pendingCalendarDelete = msg.id
-		m.pendingCalendarDeleteName = msg.name
-		message := fmt.Sprintf("Delete calendar %q?", msg.name)
-		if msg.eventCount > 0 {
-			if msg.eventCount == 1 {
-				message = fmt.Sprintf("Delete calendar %q?\n\n%d event will be deleted", msg.name, msg.eventCount)
-			} else {
-				message = fmt.Sprintf("Delete calendar %q?\n\n%d events will be deleted", msg.name, msg.eventCount)
-			}
-		}
-		if m.pendingCalendarPromoteName != "" {
-			message += fmt.Sprintf("\n\n%q will become the default.", m.pendingCalendarPromoteName)
-		}
-		m.confirmDialog = NewConfirmDialogModel(message, "Delete", m.theme).
-			Destructive().
-			SetSize(m.width, m.height)
-		m.confirmOpen = true
-		return m, nil
-
-	case CalendarManagerClosedMsg:
-		m.calendarManagerOpen = false
-		if m.pendingAccountManagementID != 0 {
-			m.accountManagementGeneration++
-			m.pendingAccountManagementID = 0
-			m.syncing = false
-			m.syncStatus = ""
-		}
-		m.pendingAccountSelection = nil
-		m.pendingAccountDefaultCandidates = nil
-		if m.pendingScopeKind == pendingScopeAccountSelectionPromote {
-			m.pendingScopeKind = pendingScopeNone
-		}
-		return m, nil
-
-	case HelpDialogRequestedMsg:
-		m.helpDialog = NewHelpDialogModel(m.theme).SetSize(m.width, m.height)
-		m.helpDialogOpen = true
-		return m, nil
-
-	case TrashViewRequestedMsg:
-		if m.trashOpen {
-			return m, nil
-		}
-		m.trashOpen = true
-		m.trash = NewTrashModel(m.calendars, newThemedHelp(m.theme)).
-			SetSelectedColor(m.theme.Selected).
-			SetSize(m.width, m.height)
-		return m, m.loadTrash()
-
-	case HelpDialogClosedMsg:
-		m.helpDialogOpen = false
-		return m, nil
-
-	case SyncAllRequestedMsg:
-		if m.syncing {
-			return m, nil
-		}
-		m.syncing = true
-		m.statusToken++
-		m.syncStatus = "Preparing sync…"
-		m.syncTargets = nil
-		m.syncTotals = syncTotals{}
-		return m, tea.Batch(m.runSyncAllPlan(), m.syncSpinner.Tick)
-
-	case syncAllPlannedMsg:
-		if len(msg.targets) == 0 {
-			m.syncing = false
-			m.statusToken++
-			m.syncStatus = "No connected calendars to sync"
-			return m, m.expireStatusAfter(6*time.Second, m.statusToken)
-		}
-		m.syncTargets = msg.targets
-		m.syncTotals = syncTotals{}
-		first := msg.targets[0]
-		m.statusToken++
-		m.syncStatus = fmt.Sprintf("Syncing %s (1/%d)…", syncProgressLabel(first.Name), len(msg.targets))
-		return m, m.runSyncOne(first, 0, len(msg.targets))
-
-	case syncCalendarFinishedMsg:
-		if msg.result != nil {
-			m.syncTotals.pushed += msg.result.Pushed
-			m.syncTotals.pulled += msg.result.Pulled
-			m.syncTotals.deleted += msg.result.Deleted
-			m.syncTotals.conflicts += msg.result.Conflicts
-			m.syncTotals.autoResolved += msg.result.AutoResolved
-			m.syncTotals.skippedConflicts += msg.result.SkippedConflicts
-			m.syncTotals.warnings += len(msg.result.Warnings)
-			m.syncTotals.errCount += len(msg.result.Errors)
-			if m.syncTotals.firstErr == nil && len(msg.result.Errors) > 0 {
-				m.syncTotals.firstErr = msg.result.Errors[0]
-			}
-		}
-		if msg.err != nil {
-			m.syncTotals.errCount++
-			if m.syncTotals.firstErr == nil {
-				m.syncTotals.firstErr = msg.err
-			}
-		}
-		next := msg.index + 1
-		if next >= msg.total {
-			label := fmt.Sprintf("%d calendar(s)", msg.total)
-			summary := syncSummary(label, m.syncTotals)
-			finishErr := m.syncTotals.firstErr
-			m.syncTargets = nil
-			m.syncTotals = syncTotals{}
-			return m, func() tea.Msg {
-				return syncFinishedMsg{summary: summary, err: finishErr, reload: true}
-			}
-		}
-		target := m.syncTargets[next]
-		m.statusToken++
-		m.syncStatus = fmt.Sprintf("Syncing %s (%d/%d)…", syncProgressLabel(target.Name), next+1, msg.total)
-		return m, m.runSyncOne(target, next, msg.total)
-
-	case SyncCalendarRequestedMsg:
-		if m.syncing {
-			// A sync is already running (e.g. re-auth finished mid-sync).
-			// Queue this one so it isn't lost — syncFinishedMsg drains it,
-			// guaranteeing the post-reauth sync runs and the ⚠ clears.
-			m.pendingSyncCalendar = syncTarget(msg)
-			return m, nil
-		}
-		m.syncing = true
-		m.statusToken++
-		label := msg.Name
-		if label == "" {
-			label = "calendar"
-		}
-		m.syncStatus = fmt.Sprintf("Syncing %s…", label)
-		return m, tea.Batch(m.runSyncCalendar(msg.ID, msg.Name), m.syncSpinner.Tick)
-
-	case spinner.TickMsg:
-		// Multiple spinners can be live simultaneously: the sync footer's, the
-		// OAuth pending modal's, and the palette's search spinner. Each bubbles
-		// spinner ignores ticks that aren't its own (ID check), so routing to
-		// all active sub-components is safe. Spinner ticks are whitelisted
-		// through every overlay guard so they always reach this handler rather
-		// than being silently consumed by the overlay.
-		var cmds []tea.Cmd
-		if m.oauthFlowOpen {
-			var c tea.Cmd
-			m.oauthFlow, c = m.oauthFlow.Update(msg)
-			if c != nil {
-				cmds = append(cmds, c)
-			}
-		}
-		if m.paletteOpen {
-			var c tea.Cmd
-			m.palette, c = m.palette.Update(msg)
-			if c != nil {
-				cmds = append(cmds, c)
-			}
-		}
-		if m.syncing {
-			var c tea.Cmd
-			m.syncSpinner, c = m.syncSpinner.Update(msg)
-			if c != nil {
-				cmds = append(cmds, c)
-			}
-		}
-		if len(cmds) == 0 {
-			return m, nil
-		}
-		return m, tea.Batch(cmds...)
-
-	case opportunisticPushFinishedMsg:
-		// Don't stomp the manual-sync status line or reset m.syncing.
-		if m.syncing {
-			return m, nil
-		}
-		if msg.err == nil && msg.summary == "" {
-			return m, nil
-		}
-		m.statusToken++
-		if msg.err != nil {
-			if msg.summary != "" {
-				m.syncStatus = fmt.Sprintf("%s — %s", msg.summary, msg.err.Error())
-			} else {
-				m.syncStatus = "Sync failed: " + msg.err.Error()
-			}
-		} else {
-			m.syncStatus = msg.summary
-		}
-		return m, m.expireStatusAfter(4*time.Second, m.statusToken)
-
-	case syncStatusExpiredMsg:
-		if msg.token == m.statusToken && !m.syncing {
-			m.syncStatus = ""
-		}
-		return m, nil
-
-	case toastTickMsg:
-		m.toast.Update(msg)
-		return m, nil
-
-	case calendarMutationDoneMsg:
-		if m.calendarManagerOpen {
-			m.calendarManager, _ = m.calendarManager.Update(msg)
-		}
-		if msg.err != nil {
-			if m.calendarManagerOpen {
-				m.calendarManager = m.calendarManager.FormSetError(cdIdxName, msg.err.Error())
-				return m, nil
-			}
-			m.err = msg.err
-			return m, nil
-		}
-		// m.calendarManagerOpen = false (Removed to keep it open)
-		return m, tea.Batch(m.loadCalendars(), m.loadEvents())
-
-	case ChoiceDialogResultMsg:
-		m.choiceOpen = false
-		kind := m.pendingScopeKind
-		m.pendingScopeKind = pendingScopeNone
-		if msg.Choice < 0 {
-			m.pendingEditSave = EventFormSaveMsg{}
-			m.pendingDelete = event.Event{}
-			m.viewReturnEvent = event.Event{}
-			if kind == pendingScopeCalendarPromote {
-				m.pendingCalendarDelete = 0
-				m.pendingCalendarDeleteName = ""
-				m.pendingCalendarPromote = 0
-				m.pendingCalendarPromoteName = ""
-				m.pendingCalendarPromoteCands = nil
-			}
-			if kind == pendingScopeAccountSelectionPromote {
-				m.pendingAccountSelection = nil
-				m.pendingAccountDefaultCandidates = nil
-			}
-			if kind == pendingScopeCalendarMoveAccount || kind == pendingScopeCalendarMoveCollection {
-				m.pendingCalendarMove = nil
-			}
-			return m, nil
-		}
-		if next, cmd, handled := m.handleCalendarMoveChoice(kind, msg.Choice); handled {
-			return next, cmd
-		}
-		if kind == pendingScopeAccountSelectionPromote {
-			if msg.Choice >= len(m.pendingAccountDefaultCandidates) ||
-				m.pendingAccountSelection == nil {
-				m.pendingAccountSelection = nil
-				m.pendingAccountDefaultCandidates = nil
-				return m, nil
-			}
-			candidate := m.pendingAccountDefaultCandidates[msg.Choice]
-			if candidate.id != 0 {
-				m.pendingAccountSelection.params.NewDefaultID = candidate.id
-			} else {
-				m.pendingAccountSelection.params.NewDefaultPath = candidate.path
-			}
-			selection := m.pendingAccountSelection
-			m.pendingAccountDefaultCandidates = nil
-			m = m.showAccountCalendarRemovalConfirmation(selection)
-			return m, nil
-		}
-		if kind == pendingScopeCalendarPromote {
-			if msg.Choice < 0 || msg.Choice >= len(m.pendingCalendarPromoteCands) {
-				m.pendingCalendarDelete = 0
-				m.pendingCalendarDeleteName = ""
-				m.pendingCalendarPromote = 0
-				m.pendingCalendarPromoteName = ""
-				m.pendingCalendarPromoteCands = nil
-				return m, nil
-			}
-			promoteID := m.pendingCalendarPromoteCands[msg.Choice]
-			m.pendingCalendarPromote = promoteID
-			if info, ok := m.calendars[promoteID]; ok {
-				m.pendingCalendarPromoteName = info.Name
-			}
-			id, name := m.pendingCalendarDelete, m.pendingCalendarDeleteName
-			m.pendingCalendarPromoteCands = nil
-			return m, func() tea.Msg {
-				count, _ := m.app.Events.CountByCalendar(context.Background(), id)
-				return calendarDeleteCountMsg{id: id, name: name, eventCount: count}
-			}
-		}
-		if kind == pendingScopeEdit {
-			save := m.pendingEditSave
-			m.pendingEditSave = EventFormSaveMsg{}
-			editID := m.form.editID
-			choice := msg.Choice
-			return m, func() tea.Msg {
-				return m.dispatchEditScope(editID, choice, save)
-			}
-		}
-		ev := m.pendingDelete
-		return m, func() tea.Msg {
-			switch msg.Choice {
-			case 0: // This event
-				meta, err := m.app.Events.DeleteInstanceWithUndo(context.Background(), ev.UID, ev.StartTime)
-				return eventDeletedMsg{
-					calendarID: ev.CalendarID,
-					meta:       meta,
-					title:      ev.Title,
-					err:        err,
-				}
-			case 1: // This and following
-				meta, err := m.app.Events.DeleteFromInstanceWithUndo(context.Background(), ev.UID, ev.StartTime)
-				return eventDeletedMsg{
-					calendarID: ev.CalendarID,
-					meta:       meta,
-					title:      ev.Title,
-					err:        err,
-				}
-			case 2: // All events
-				meta, err := m.app.Events.DeleteSeriesWithUndo(context.Background(), ev.UID)
-				return eventDeletedMsg{
-					calendarID: ev.CalendarID,
-					meta:       meta,
-					title:      ev.Title,
-					err:        err,
-				}
-			}
-			return eventDeletedMsg{calendarID: ev.CalendarID}
-		}
-
-	case ConfirmDialogResultMsg:
-		m.confirmOpen = false
-		if m.pendingQuit {
-			m.pendingQuit = false
-			if msg.Confirmed {
-				m.oauthFlow.Abort() // release any in-flight OAuth listener
-				return m, tea.Quit
-			}
-			return m, nil
-		}
-		if !msg.Confirmed {
-			m.pendingCalendarDelete = 0
-			m.pendingCalendarDeleteName = ""
-			m.pendingCalendarKeepLocal = 0
-			m.pendingCalendarPromote = 0
-			m.pendingCalendarPromoteName = ""
-			m.pendingPurgeEntries = nil
-			m.pendingPurgeTitle = ""
-			m.pendingAccountSelection = nil
-			m.pendingAccountDefaultCandidates = nil
-			m.pendingAccountRemoveID = 0
-			m.pendingAccountRemoveName = ""
-			return m, nil
-		}
-		if m.pendingAccountRemoveID != 0 {
-			accountID := m.pendingAccountRemoveID
-			name := m.pendingAccountRemoveName
-			m.pendingAccountRemoveID = 0
-			m.pendingAccountRemoveName = ""
-			m.calendarManagerOpen = false
-			m.syncing = true
-			m.syncStatus = "Removing account…"
-			return m, tea.Batch(m.syncSpinner.Tick, m.removeAccount(accountID, name))
-		}
-		if m.pendingAccountSelection != nil {
-			selection := m.pendingAccountSelection
-			m.pendingAccountSelection = nil
-			m.pendingAccountDefaultCandidates = nil
-			m.syncing = true
-			m.syncStatus = "Applying calendar changes…"
-			return m, tea.Batch(
-				m.syncSpinner.Tick,
-				m.reconcileAndSyncAccountCalendars(selection),
-			)
-		}
-		if len(m.pendingPurgeEntries) > 0 {
-			entries := m.pendingPurgeEntries
-			title := m.pendingPurgeTitle
-			m.pendingPurgeEntries = nil
-			m.pendingPurgeTitle = ""
-			return m, func() tea.Msg {
-				for _, e := range entries {
-					if err := m.app.Trash.Purge(context.Background(), e); err != nil {
-						return trashActionDoneMsg{action: "purged", title: title, err: err}
-					}
-				}
-				return trashActionDoneMsg{action: "purged", title: title, err: nil}
-			}
-		}
-		if m.pendingCalendarKeepLocal != 0 {
-			id := m.pendingCalendarKeepLocal
-			m.pendingCalendarKeepLocal = 0
-			return m, func() tea.Msg {
-				ctx := context.Background()
-				cal, err := m.app.Calendars.Get(ctx, id)
-				if err != nil {
-					return calendarMutationDoneMsg{err: err}
-				}
-				credStore, _ := m.openCredentialStore()
-				return calendarMutationDoneMsg{err: m.app.Calendars.Disconnect(ctx, cal, credStore)}
-			}
-		}
-		if m.pendingCalendarDelete != 0 {
-			id := m.pendingCalendarDelete
-			newDefaultID := m.pendingCalendarPromote
-			m.pendingCalendarDelete = 0
-			m.pendingCalendarDeleteName = ""
-			m.pendingCalendarPromote = 0
-			m.pendingCalendarPromoteName = ""
-			// Delete confirmed: close the edit dialog too.
-			m.calendarManagerOpen = false
-			return m, func() tea.Msg {
-				credStore, _ := m.openCredentialStore()
-				err := m.app.Calendars.DeleteWithRemoteCleanup(context.Background(), id, newDefaultID, credStore)
-				return calendarMutationDoneMsg{err: err}
-			}
-		}
-		ev := m.pendingDelete
-		return m, func() tea.Msg {
-			meta, err := m.app.Events.DeleteWithUndo(context.Background(), ev.ID)
-			return eventDeletedMsg{
-				calendarID: ev.CalendarID,
-				meta:       meta,
-				title:      ev.Title,
-				err:        err,
-			}
-		}
-
-	case eventDeletedMsg:
-		if msg.err != nil {
-			m.err = msg.err
-			return m, nil
-		}
-		m.viewDialogOpen = false
-		// Every soft-delete is reversible; push the undo entry unconditionally.
-		// UID being empty means no row was actually deleted (defensive).
-		if msg.meta.UID == "" {
-			return m, tea.Batch(m.loadEvents(), m.runOpportunisticPush(msg.calendarID))
-		}
-		m.undoStack.Push(UndoEntry{
-			Meta:      msg.meta,
-			DeletedAt: time.Now(),
-		})
-		synced := false // opportunistic push hasn't run yet when toast shows
-		toastCmd := m.toast.Deleted(msg.title, synced)
-		m.pushDeferralToken++
-		token := m.pushDeferralToken
-		calID := msg.calendarID
-		deferCmd := tea.Tick(ToastAutoDismissDelay, func(time.Time) tea.Msg {
-			return deferredPushMsg{calendarID: calID, token: token}
-		})
-		return m, tea.Batch(m.loadEvents(), toastCmd, deferCmd)
-
-	case deferredPushMsg:
-		// If a restore bumped the token between the delete and this tick,
-		// the deferred push is stale — drop it.
-		if msg.token != m.pushDeferralToken {
-			return m, nil
-		}
-		if push := m.runOpportunisticPush(msg.calendarID); push != nil {
-			return m, push
-		}
-		return m, nil
-
-	case eventRestoredMsg:
-		// The restore has landed; allow the next undo to dispatch.
-		m.undoRestoreInFlight = false
-		if msg.err != nil {
-			// Route the dismiss tick — previously dropped, so failed toasts
-			// never auto-cleared in the live app.
-			cmd := m.toast.Failed(msg.err.Error())
-			// Leave the entry on the stack so the user can retry with
-			// different context (e.g. restoring the calendar first).
-			return m, cmd
-		}
-		// Remove the entry that was actually restored, by identity, not the
-		// current top. A delete that lands while the restore is in flight may
-		// have pushed a newer entry. That entry's undo affordance must survive.
-		m.undoStack.Remove(msg.meta)
-		toastCmd := m.toast.Restored(msg.title)
-		return m, tea.Batch(m.loadEvents(), toastCmd)
-
-	case TrashDialogClosedMsg:
-		m.trashOpen = false
-		return m, nil
-
-	case trashLoadedMsg:
-		if msg.err != nil {
-			m.err = msg.err
-			return m, nil
-		}
-		m.trash = m.trash.SetEntries(msg.entries, m.calendars)
-		return m, nil
-
-	case TrashReloadMsg:
-		return m, m.loadTrash()
-
-	case TrashRestoreRequestedMsg:
-		entries := msg.Entries
-		if len(entries) == 0 {
-			return m, nil
-		}
-		title := trashBulkTitle(entries)
-		return m, func() tea.Msg {
-			for _, e := range entries {
-				if err := m.app.Trash.Restore(context.Background(), e); err != nil {
-					return trashActionDoneMsg{action: "restored", title: title, err: err}
-				}
-			}
-			return trashActionDoneMsg{action: "restored", title: title, err: nil}
-		}
-
-	case TrashPurgeRequestedMsg:
-		if len(msg.Entries) == 0 {
-			return m, nil
-		}
-		m.pendingPurgeEntries = msg.Entries
-		m.pendingPurgeTitle = trashBulkTitle(msg.Entries)
-		var message string
-		if len(msg.Entries) == 1 {
-			message = fmt.Sprintf("Purge %q forever? This can't be undone.", msg.Entries[0].Title)
-		} else {
-			message = fmt.Sprintf("Purge %d items forever? This can't be undone.", len(msg.Entries))
-		}
-		m.confirmDialog = NewConfirmDialogModel(message, "Purge", m.theme).
-			Destructive().
-			SetSize(m.width, m.height)
-		m.confirmOpen = true
-		return m, nil
-
-	case trashActionDoneMsg:
-		if msg.err != nil {
-			cmd := m.toast.Failed(msg.err.Error())
-			return m, cmd
-		}
-		m.trash = m.trash.ClearMarks()
-		cmds := []tea.Cmd{m.loadTrash(), m.loadEvents()}
-		switch msg.action {
-		case "restored":
-			cmds = append(cmds, m.toast.Restored(msg.title))
-		case "purged":
-			cmds = append(cmds, m.toast.Purged(msg.title))
-		}
-		return m, tea.Batch(cmds...)
-
-	case tea.MouseWheelMsg:
-		if m.helpDialogOpen {
-			var cmd tea.Cmd
-			m.helpDialog, cmd = m.helpDialog.Update(msg)
-			return m, cmd
-		}
-		if m.viewDialogOpen {
-			var cmd tea.Cmd
-			m.viewDialog, cmd = m.viewDialog.Update(msg)
-			return m, cmd
-		}
-		if m.dialogOpen {
-			var cmd tea.Cmd
-			m.dialog, cmd = m.dialog.Update(msg)
-			return m, cmd
-		}
-		if m.calendarManagerOpen {
-			var cmd tea.Cmd
-			m.calendarManager, cmd = m.calendarManager.Update(msg)
-			return m, cmd
-		}
-		if m.trashOpen {
-			var cmd tea.Cmd
-			m.trash, cmd = m.trash.Update(msg)
-			return m, cmd
-		}
-		if !m.dialogOpen && !m.choiceOpen && !m.confirmOpen && !m.oauthFlowOpen {
-			switch m.viewMode {
-			case viewWeek:
-				switch msg.Button {
-				case tea.MouseWheelUp:
-					m.week.scrollOffset -= m.week.linesPerHour
-					if m.week.scrollOffset < 0 {
-						m.week.scrollOffset = 0
-					}
-				case tea.MouseWheelDown:
-					m.week.scrollOffset += m.week.linesPerHour
-					if ms := m.week.maxScroll(); m.week.scrollOffset > ms {
-						m.week.scrollOffset = ms
-					}
-				}
-			case viewDay:
-				switch msg.Button {
-				case tea.MouseWheelUp:
-					m.day.scrollOffset -= m.day.linesPerHour
-					if m.day.scrollOffset < 0 {
-						m.day.scrollOffset = 0
-					}
-				case tea.MouseWheelDown:
-					m.day.scrollOffset += m.day.linesPerHour
-					if ms := m.day.maxScroll(); m.day.scrollOffset > ms {
-						m.day.scrollOffset = ms
-					}
-				}
-			case viewAgenda:
-				var cmd tea.Cmd
-				switch msg.Button {
-				case tea.MouseWheelUp:
-					cmd = m.agenda.ScrollBy(-agendaWheelStep)
-				case tea.MouseWheelDown:
-					cmd = m.agenda.ScrollBy(agendaWheelStep)
-				}
-				return m, cmd
-			default:
-				// viewMonth: no wheel scrolling
-			}
-		}
-		return m, nil
-
-	case tea.MouseClickMsg:
-		if msg.Button != tea.MouseLeft {
-			return m, nil
-		}
-		if m.helpDialogOpen {
-			var cmd tea.Cmd
-			m.helpDialog, cmd = m.helpDialog.Update(msg)
-			return m, cmd
-		}
-		if m.choiceOpen {
-			var cmd tea.Cmd
-			m.choiceDialog, cmd = m.choiceDialog.Update(msg)
-			return m, cmd
-		}
-		if m.confirmOpen {
-			var cmd tea.Cmd
-			m.confirmDialog, cmd = m.confirmDialog.Update(msg)
-			return m, cmd
-		}
-		if m.dialogOpen {
-			var cmd tea.Cmd
-			m.dialog, cmd = m.dialog.Update(msg)
-			return m, cmd
-		}
-		if m.viewDialogOpen {
-			var cmd tea.Cmd
-			m.viewDialog, cmd = m.viewDialog.Update(msg)
-			return m, cmd
-		}
-		if m.calendarManagerOpen {
-			var cmd tea.Cmd
-			m.calendarManager, cmd = m.calendarManager.Update(msg)
-			return m, cmd
-		}
-		if m.trashOpen {
-			var cmd tea.Cmd
-			m.trash, cmd = m.trash.Update(msg)
-			return m, cmd
-		}
-		// Footer label hit-test: clicking the context label (MONTH/WEEK/DAY/
-		// AGENDA) cycles through views. The sweep in View() populates the
-		// tracker with the footer label's zone whenever no overlay is open.
-		if target := MouseResolve(msg.X, msg.Y); target == "footer:label" {
-			return m.cycleView()
-		}
-		// Sidebar hit-test. The sidebar content starts at (padding, padding)
-		// inside the outer screen, with a 1-col right border. If the click
-		// lands inside that x-range we dispatch to the sidebar in its local
-		// coordinates instead of the main calendar.
-		if m.showSidebar {
-			padding := 1
-			if msg.X >= padding && msg.X < sidebarWidth-padding {
-				localX := msg.X - padding
-				localY := msg.Y - padding
-				// Moving focus to the sidebar mirrors keyboard navigation;
-				// otherwise the chevrons would click but not visibly focus.
-				if m.focus != focusSidebar {
-					m.focus = focusSidebar
-					m.sidebar = m.sidebar.Focus()
-				}
-				var cmd tea.Cmd
-				m.sidebar, cmd = m.sidebar.HandleClick(localX, localY)
-				return m, cmd
-			}
-		}
-		// Click landed outside the sidebar. Pull focus back to the main
-		// view. Later keystrokes then target the calendar, not the
-		// sidebar that was last clicked.
-		if m.focus == focusSidebar {
-			m.sidebar = m.sidebar.Blur()
-			m.focus = focusCalendar
-		}
-		ox, oy := m.calendarOffset()
-		switch m.viewMode {
-		case viewDay:
-			day, ok := m.day.DayAtPosition(msg.X-ox, msg.Y-oy)
-			if !ok {
-				return m, nil
-			}
-			m.clickedEventID = m.day.EventAtPosition(msg.X-ox, msg.Y-oy)
-			var cmd tea.Cmd
-			m.day, cmd = m.day.selectDay(day)
-			return m, cmd
-		case viewWeek:
-			day, ok := m.week.DayAtPosition(msg.X-ox, msg.Y-oy)
-			if !ok {
-				return m, nil
-			}
-			m.clickedEventID = m.week.EventAtPosition(msg.X-ox, msg.Y-oy)
-			var cmd tea.Cmd
-			m.week, cmd = m.week.selectDay(day)
-			return m, cmd
-		case viewAgenda:
-			var cmd tea.Cmd
-			m.agenda, cmd = m.agenda.HandleClick(msg.X-ox, msg.Y-oy)
-			return m, cmd
-		default:
-			day, ok := m.calendar.DayAtPosition(msg.X-ox, msg.Y-oy)
-			if !ok {
-				return m, nil
-			}
-			m.clickedEventID = m.calendar.EventAtPosition(msg.X-ox, msg.Y-oy)
-			var cmd tea.Cmd
-			m.calendar, cmd = m.calendar.selectDay(day)
-			return m, cmd
-		}
-
-	case tea.KeyPressMsg:
-		// Help sits on top of every other overlay (see View()), so it must
-		// own input — including Esc — whenever it's open.
-		if m.helpDialogOpen {
-			var cmd tea.Cmd
-			m.helpDialog, cmd = m.helpDialog.Update(msg)
-			return m, cmd
-		}
-		if m.choiceOpen {
-			var cmd tea.Cmd
-			m.choiceDialog, cmd = m.choiceDialog.Update(msg)
-			return m, cmd
-		}
-		if m.confirmOpen {
-			var cmd tea.Cmd
-			m.confirmDialog, cmd = m.confirmDialog.Update(msg)
-			return m, cmd
-		}
-		if m.viewDialogOpen {
-			var cmd tea.Cmd
-			m.viewDialog, cmd = m.viewDialog.Update(msg)
-			return m, cmd
-		}
-		if m.dialogOpen {
-			var cmd tea.Cmd
-			m.dialog, cmd = m.dialog.Update(msg)
-			return m, cmd
-		}
-		if m.calendarManagerOpen {
-			var cmd tea.Cmd
-			m.calendarManager, cmd = m.calendarManager.Update(msg)
-			return m, cmd
-		}
-		if m.trashOpen {
-			var cmd tea.Cmd
-			m.trash, cmd = m.trash.Update(msg)
-			return m, cmd
-		}
-		switch {
-		case key.Matches(msg, m.keys.Palette):
-			return m.openPalette()
-		case key.Matches(msg, m.keys.MonthView):
-			return m.switchToView(viewMonth)
-		case key.Matches(msg, m.keys.WeekView):
-			return m.switchToView(viewWeek)
-		case key.Matches(msg, m.keys.DayView):
-			return m.switchToView(viewDay)
-		case key.Matches(msg, m.keys.AgendaView):
-			return m.switchToView(viewAgenda)
-		case key.Matches(msg, m.keys.Sidebar):
-			return m.toggleSidebar()
-		case key.Matches(msg, m.keys.WeekNumbers):
-			return m.toggleWeekNumbers()
-		case key.Matches(msg, m.keys.WeekStart):
-			return m.toggleWeekStart()
-		case key.Matches(msg, m.keys.CalendarList):
-			return m, func() tea.Msg { return CalendarManagerRequestedMsg{Target: CalendarManagerTargetRoot} }
-		case key.Matches(msg, m.keys.Sync):
-			return m, func() tea.Msg { return SyncAllRequestedMsg{} }
-		case key.Matches(msg, m.keys.SwitchFocus):
-			// Only handle Tab/Shift+Tab at the app level when the user
-			// enters the sidebar from the main view. Forward Tab lands on
-			// the first sidebar tab stop (the prev-month chevron). Backward
-			// Shift+Tab lands on the last (the bottom calendar list row).
-			// Once focus is inside the sidebar, the key falls through to
-			// m.sidebar.Update. That cycles between its internal stops and
-			// emits SidebarFocusEscapedMsg to hand focus back to the main
-			// view.
-			if m.showSidebar && m.focus != focusSidebar {
-				m.focus = focusSidebar
-				if msg.String() == "shift+tab" {
-					m.sidebar = m.sidebar.FocusAtEnd()
-				} else {
-					m.sidebar = m.sidebar.FocusAtStart()
-				}
-				return m, nil
-			}
-			// Fall through to the sidebar routing below.
-		case key.Matches(msg, m.keys.Create):
-			var cursor time.Time
-			switch m.viewMode {
-			case viewDay:
-				cursor = m.day.Cursor()
-			case viewWeek:
-				cursor = m.week.Cursor()
-			case viewAgenda:
-				cursor = m.agenda.SelectedDay()
-			default:
-				cursor = m.calendar.Cursor()
-			}
-			form, cmd := NewEventFormModel(cursor, eventFormCalendars(m.calendars), m.theme)
-			return m.openEventForm(form, cmd)
-		}
-		if m.focus == focusCalendar {
-			switch m.viewMode {
-			case viewDay:
-				var cmd tea.Cmd
-				m.day, cmd = m.day.Update(msg)
-				return m, cmd
-			case viewWeek:
-				var cmd tea.Cmd
-				m.week, cmd = m.week.Update(msg)
-				return m, cmd
-			case viewAgenda:
-				var cmd tea.Cmd
-				m.agenda, cmd = m.agenda.Update(msg)
-				return m, cmd
-			default:
-				var cmd tea.Cmd
-				m.calendar, cmd = m.calendar.Update(msg)
-				return m, cmd
-			}
-		}
-		if m.focus == focusSidebar {
-			var cmd tea.Cmd
-			m.sidebar, cmd = m.sidebar.Update(msg)
-			return m, cmd
-		}
-		return m, nil
-	}
-
-	return m, nil
-}
-
-func (m Model) View() tea.View {
-	v := tea.View{AltScreen: true, MouseMode: tea.MouseModeCellMotion}
-
-	if m.width == 0 {
-		return v
-	}
-
-	if !m.ready {
-		v.Content = "\n  Loading..."
-		return v
-	}
-
-	padding := 1
-	mainWidth, contentHeight := m.mainDims()
-
-	var mainContent string
-	switch m.viewMode {
-	case viewDay:
-		mainContent = m.day.View()
-	case viewWeek:
-		mainContent = m.week.View()
-	case viewAgenda:
-		mainContent = m.agenda.View()
-	default:
-		mainContent = m.calendar.View()
-	}
-	if m.err != nil {
-		mainContent = lipgloss.NewStyle().Foreground(m.theme.Error).Render("Error: " + m.err.Error())
-	}
-
-	main := lipgloss.NewStyle().
-		Width(mainWidth).
-		Height(contentHeight).
-		Padding(padding).
-		Foreground(m.theme.Text).
-		Render(mainContent)
-
-	var body string
-	if m.showSidebar {
-		// Render content and border as two siblings joined horizontally.
-		// Faint (used to match the calendar grid chrome when unfocused)
-		// then applies to the border glyphs only. It never dims the
-		// sidebar's inner text.
-		sb := m.sidebar.SetSize(sidebarWidth-padding*2, contentHeight-padding*2)
-		inner := lipgloss.NewStyle().
-			Width(sidebarWidth).
-			Height(contentHeight).
-			Padding(padding).
-			Foreground(m.theme.Text).
-			Render(sb.View())
-
-		borderStyle := lipgloss.NewStyle().Faint(true)
-		if m.focus == focusSidebar {
-			borderStyle = lipgloss.NewStyle().Foreground(m.theme.Primary)
-		}
-		borderChar := borderStyle.Render("│")
-		borderLines := make([]string, contentHeight)
-		for i := range borderLines {
-			borderLines[i] = borderChar
-		}
-		border := strings.Join(borderLines, "\n")
-
-		sidebar := lipgloss.JoinHorizontal(lipgloss.Top, inner, border)
-		body = lipgloss.JoinHorizontal(lipgloss.Top, sidebar, main)
-	} else {
-		body = main
-	}
-
-	var statusText string
-	if m.syncStatus != "" {
-		statusColor := m.theme.Primary
-		if m.syncing {
-			statusColor = m.theme.Muted
-		}
-		statusText = lipgloss.NewStyle().Foreground(statusColor).Render(m.syncStatus)
-		if m.syncing {
-			m.syncSpinner.Style = lipgloss.NewStyle().Foreground(m.theme.TextDim)
-			statusText = m.syncSpinner.View() + " " + statusText
-		}
-	}
-	innerWidth := m.width - padding*2
-	var footerLine string
-	if m.anyOverlayOpen() {
-		// A dialog owns the bottom of the screen; don't duplicate its hints.
-		// "? help" is misleading while the help dialog itself is up.
-		footerLine = m.footer.RenderMinimal(innerWidth, statusText, m.toast.View(), !m.helpDialogOpen)
-	} else {
-		footerLine = m.footer.Render(
-			m.currentFooterContext(),
-			innerWidth,
-			statusText,
-			m.toast.View(),
-			m.currentFooterHasRSVP(),
-			m.currentFooterShowsTodayHint(),
-		)
-	}
-	footer := lipgloss.NewStyle().
-		PaddingLeft(padding).
-		PaddingRight(padding).
-		Render(footerLine)
-	v.Content = lipgloss.JoinVertical(lipgloss.Left, body, footer)
-	// Strip mouse-track markers embedded by the footer label so they do not
-	// leak into terminal output. Populate the tracker's zones so clicks
-	// on the label can be resolved. Overlays, if any, will overwrite zones
-	// during their own View() sweeps below. That is fine. The footer
-	// label is not clickable while an overlay owns input.
-	v.Content = MouseSweep(v.Content)
-
-	if m.dialogOpen {
-		v.Content = m.compositeDialog(v.Content)
-	}
-	if m.viewDialogOpen {
-		bw, bh := m.viewDialog.BoxSize()
-		v.Content = m.compositeOverlay(v.Content, m.viewDialog.View(), bw, bh)
-	}
-	if m.formOpen {
-		bw, bh := m.form.BoxSize()
-		v.Content = m.compositeOverlay(v.Content, m.form.View(), bw, bh)
-		if m.form.DatePickerOpen() {
-			pw, ph := m.form.DatePickerBoxSize()
-			v.Content = m.compositeOverlay(v.Content, m.form.DatePickerView(), pw, ph)
-		}
-		if m.form.EndsDatePickerOpen() {
-			pw, ph := m.form.DatePickerBoxSize()
-			v.Content = m.compositeOverlay(v.Content, m.form.EndsDatePickerView(), pw, ph)
-		}
-		if m.form.TimezonePickerOpen() {
-			tw, th := m.form.TimezonePickerBoxSize()
-			v.Content = m.compositeOverlay(v.Content, m.form.TimezonePickerView(), tw, th)
-		}
-		if m.form.RRuleEditorOpen() {
-			ew, eh := m.form.rruleEditor.BoxSize()
-			v.Content = m.compositeOverlay(v.Content, m.form.rruleEditor.View(), ew, eh)
-			if m.form.rruleEditor.EndsDatePickerOpen() {
-				pw, ph := m.form.rruleEditor.EndsDatePickerBoxSize()
-				v.Content = m.compositeOverlay(v.Content, m.form.rruleEditor.EndsDatePickerView(), pw, ph)
-			}
-		}
-		if m.form.AlarmEditorOpen() {
-			ew, eh := m.form.alarmEditor.BoxSize()
-			v.Content = m.compositeOverlay(v.Content, m.form.alarmEditor.View(), ew, eh)
-		}
-	}
-	if m.calendarManagerOpen {
-		bw, bh := m.calendarManager.BoxSize()
-		v.Content = m.compositeOverlay(v.Content, m.calendarManager.View(), bw, bh)
-	}
-	if m.trashOpen {
-		bw, bh := m.trash.BoxSize()
-		v.Content = m.compositeOverlay(v.Content, m.trash.View(), bw, bh)
-	}
-	if m.accountRenameOpen {
-		bw, bh := m.accountRename.BoxSize()
-		v.Content = m.compositeOverlay(v.Content, m.accountRename.View(), bw, bh)
-	}
-	if m.accountOAuthConfigOpen {
-		bw, bh := m.accountOAuthConfig.BoxSize()
-		v.Content = m.compositeOverlay(v.Content, m.accountOAuthConfig.View(), bw, bh)
-	}
-	if m.accountCredentialsOpen {
-		bw, bh := m.accountCredentials.BoxSize()
-		v.Content = m.compositeOverlay(v.Content, m.accountCredentials.View(), bw, bh)
-	}
-	if m.oauthFlowOpen {
-		bw, bh := m.oauthFlow.BoxSize()
-		v.Content = m.compositeOverlay(v.Content, m.oauthFlow.View(), bw, bh)
-	}
-	if m.choiceOpen {
-		bw, bh := m.choiceDialog.BoxSize()
-		v.Content = m.compositeOverlay(v.Content, m.choiceDialog.View(), bw, bh)
-	}
-	// Regular confirms belong in the normal stack. The quit guard must
-	// sit above palette and help (which otherwise render on top). It
-	// owns input whenever pendingQuit is set.
-	if m.confirmOpen && !m.pendingQuit {
-		bw, bh := m.confirmDialog.BoxSize()
-		v.Content = m.compositeOverlay(v.Content, m.confirmDialog.View(), bw, bh)
-	}
-	if m.paletteOpen {
-		bw, bh := m.palette.BoxSize()
-		v.Content = m.compositeOverlay(v.Content, m.palette.View(), bw, bh)
-	}
-	if m.helpDialogOpen {
-		bw, bh := m.helpDialog.BoxSize()
-		v.Content = m.compositeOverlay(v.Content, m.helpDialog.View(), bw, bh)
-	}
-	if m.confirmOpen && m.pendingQuit {
-		bw, bh := m.confirmDialog.BoxSize()
-		v.Content = m.compositeOverlay(v.Content, m.confirmDialog.View(), bw, bh)
-	}
-
-	return v
-}
-
-// compositeDialog draws the dialog box over the already-rendered main view
-// using an ultraviolet screen buffer. The background content outside the
-// dialog's rectangle is preserved unchanged.
-func (m Model) compositeDialog(background string) string {
-	if m.width <= 0 || m.height <= 0 {
-		return background
-	}
-
-	buf := uv.NewScreenBuffer(m.width, m.height)
-	uv.NewStyledString(background).Draw(buf, buf.Bounds())
-
-	dialogView := m.dialog.View()
-	if dialogView == "" {
-		return buf.Render()
-	}
-
-	boxW, boxH := m.dialog.BoxSize()
-	if boxW <= 0 || boxH <= 0 {
-		return buf.Render()
-	}
-	x := (m.width - boxW) / 2
-	y := (m.height - boxH) / 2
-	rect := image.Rect(x, y, x+boxW, y+boxH)
-	uv.NewStyledString(dialogView).Draw(buf, rect)
-
-	return buf.Render()
-}
-
-func (m Model) compositeOverlay(background, overlay string, boxW, boxH int) string {
-	if m.width <= 0 || m.height <= 0 || boxW <= 0 || boxH <= 0 || overlay == "" {
-		return background
-	}
-	buf := uv.NewScreenBuffer(m.width, m.height)
-	uv.NewStyledString(background).Draw(buf, buf.Bounds())
-	x := (m.width - boxW) / 2
-	y := (m.height - boxH) / 2
-	rect := image.Rect(x, y, x+boxW, y+boxH)
-	uv.NewStyledString(overlay).Draw(buf, rect)
-	return buf.Render()
-}
-
-// viewCursorAndToday returns the cursor and today from whichever view is active.
-func (m Model) viewCursorAndToday() (time.Time, time.Time) {
-	switch m.viewMode {
-	case viewDay:
-		return m.day.cursor, m.day.today
-	case viewWeek:
-		return m.week.cursor, m.week.today
-	case viewAgenda:
-		return m.agenda.cursor, m.agenda.today
-	default:
-		return m.calendar.cursor, m.calendar.today
-	}
-}
-
-// switchToView changes the active view mode and synchronizes cursor/today.
-// Safe to call even when already in the requested mode (no-op).
-func (m Model) switchToView(mode viewMode) (tea.Model, tea.Cmd) {
-	if m.viewMode == mode {
-		var cmd tea.Cmd
-		m, cmd = m.scheduleClockTick()
-		return m, cmd
-	}
-	if m.focus == focusSidebar {
-		m.sidebar = m.sidebar.Blur()
-		m.focus = focusCalendar
-	}
-	cursor, today := m.viewCursorAndToday()
-	m.viewMode = mode
-	switch mode {
-	case viewMonth:
-		m.calendar.cursor = cursor
-		m.calendar.today = today
-		if m.calendar.cursor.Year() != m.calendar.month.Year() || m.calendar.cursor.Month() != m.calendar.month.Month() {
-			m.calendar.month = time.Date(cursor.Year(), cursor.Month(), 1, 0, 0, 0, 0, cursor.Location())
-		}
-	case viewWeek:
-		m.week.cursor = cursor
-		m.week.today = today
-	case viewDay:
-		m.day.cursor = cursor
-		m.day.today = today
-	case viewAgenda:
-		m.agenda.cursor = cursor
-		m.agenda.today = today
-		m.agenda = m.agenda.ResetWindow(cursor)
-		if sameDay(cursor, today) {
-			m.agenda = m.agenda.SelectCurrentOrNext(time.Now())
-		}
-	}
-	cmds := []tea.Cmd{m.switchView()}
-	var cmd tea.Cmd
-	m, cmd = m.scheduleClockTick()
-	if cmd != nil {
-		cmds = append(cmds, cmd)
-	}
-	return m, tea.Batch(cmds...)
-}
-
-// cycleView advances to the next view mode in the order month → week → day
-// → agenda → month. It keeps cursor/today across the switch.
-func (m Model) cycleView() (tea.Model, tea.Cmd) {
-	var next viewMode
-	switch m.viewMode {
-	case viewMonth:
-		next = viewWeek
-	case viewWeek:
-		next = viewDay
-	case viewDay:
-		next = viewAgenda
-	default:
-		next = viewMonth
-	}
-	return m.switchToView(next)
-}
-
-// goToToday moves the cursor in the active view to today and reloads events.
-func (m Model) goToToday() (tea.Model, tea.Cmd) {
-	now := time.Now()
-	today := time.Date(now.Year(), now.Month(), now.Day(), 0, 0, 0, 0, now.Location())
-	switch m.viewMode {
-	case viewDay:
-		m.day.cursor = today
-		m.day.today = today
-	case viewWeek:
-		m.week.cursor = today
-		m.week.today = today
-	case viewAgenda:
-		m.agenda.cursor = today
-		m.agenda.today = today
-		m.agenda = m.agenda.ResetWindow(today)
-	default:
-		m.calendar.cursor = today
-		m.calendar.today = today
-		if m.calendar.month.Year() != today.Year() || m.calendar.month.Month() != today.Month() {
-			m.calendar.month = time.Date(today.Year(), today.Month(), 1, 0, 0, 0, 0, today.Location())
-		}
-	}
-	return m, m.loadEvents()
-}
-
-// toggleWeekNumbers toggles the ISO week-number gutter in month/week views.
-func (m Model) toggleWeekNumbers() (tea.Model, tea.Cmd) {
-	m.showWeekNumbers = !m.showWeekNumbers
-	m.calendar = m.calendar.SetShowWeekNumbers(m.showWeekNumbers)
-	m.week = m.week.SetShowWeekNumbers(m.showWeekNumbers)
-	m.saveUIState()
-	return m, nil
-}
-
-// toggleWeekStart switches the first day of the week between Sunday and Monday.
-// Month and week query ranges follow the new grid, so events reload.
-func (m Model) toggleWeekStart() (tea.Model, tea.Cmd) {
-	if m.weekStart == time.Monday {
-		m.weekStart = time.Sunday
-	} else {
-		m.weekStart = time.Monday
-	}
-	m.calendar = m.calendar.SetWeekStart(m.weekStart)
-	m.week = m.week.SetWeekStart(m.weekStart)
-	m.sidebar = m.sidebar.SetWeekStart(m.weekStart)
-	m.saveUIState()
-	return m, m.loadEvents()
-}
-
-// toggleSidebar toggles the sidebar panel and resyncs view sizes.
-func (m Model) toggleSidebar() (tea.Model, tea.Cmd) {
-	m.showSidebar = !m.showSidebar
-	if !m.showSidebar {
-		m.focus = focusCalendar
-	}
-	iw, ih := m.innerDims()
-	m.calendar = m.calendar.SetSize(iw, ih)
-	m.week = m.week.SetSize(iw, ih)
-	m.day = m.day.SetSize(iw, ih)
-	m.agenda = m.agenda.SetSize(iw, ih)
-	m.saveUIState()
-	return m, nil
-}
-
-// openEventForm mounts an event form with the session week start and size.
-func (m Model) openEventForm(form EventFormModel, cmd tea.Cmd) (Model, tea.Cmd) {
-	m.form = form.SetWeekStart(m.weekStart).SetSize(m.width, m.height)
-	m.formOpen = true
-	return m, cmd
-}
-
-// openPalette initializes and shows the command palette.
-func (m Model) openPalette() (tea.Model, tea.Cmd) {
-	cmds := buildPaletteCommands(m)
-	palette, cmd := NewPaletteModel(cmds, m.theme, makePaletteSearchFunc(m))
-	palette = palette.SetSize(m.width, m.height)
-	m.palette = palette
-	m.paletteOpen = true
-	return m, cmd
-}
-
-// switchView resizes all views and reloads events after a view mode change.
-func (m *Model) switchView() tea.Cmd {
-	iw, ih := m.innerDims()
-	m.calendar = m.calendar.SetSize(iw, ih)
-	m.week = m.week.SetSize(iw, ih)
-	m.day = m.day.SetSize(iw, ih)
-	m.agenda = m.agenda.SetSize(iw, ih)
-	m.saveUIState()
-	return m.loadEvents()
-}
-
-func (m Model) saveUIState() {
-	var vm string
-	switch m.viewMode {
-	case viewWeek:
-		vm = "week"
-	case viewDay:
-		vm = "day"
-	case viewAgenda:
-		vm = "agenda"
-	default:
-		vm = "month"
-	}
-	ids := make([]int64, 0, len(m.hiddenCalendars))
-	for id, hidden := range m.hiddenCalendars {
-		if hidden {
-			ids = append(ids, id)
-		}
-	}
-	slices.Sort(ids)
-	_ = config.SaveUIState(config.UIState{
-		ShowSidebar:         m.showSidebar,
-		ViewMode:            vm,
-		HiddenCalendars:     ids,
-		AgendaShowEmptyDays: m.agenda.ShowEmptyDays(),
-		ShowWeekNumbers:     m.showWeekNumbers,
-		WeekStart:           config.FormatWeekStart(m.weekStart),
-	})
 }
 
 // RunOptions configures a TUI session. A non-zero Event jumps to that
